### 2019-09-13 #1727 @whedon
[PRE REVIEW]: FAT Forensics: A Python Toolbox for Implementing and Deploying Fairness, Accountability and Transparency Algorithms in Predictive Systems
**Submitting author:** @So-Cool (<a href="http://orcid.org/0000-0002-9869-5896">Kacper Sokol</a>)
**Repository:** <a href="https://github.com/fat-forensics/fat-forensics" target ="_blank">https://github.com/fat-forensics/fat-forensics</a>
**Version:** 0.0.1
