### 2019-09-14 #1727 @arokem
👋 @ahurriyetoglu : would you be willing to review this submission for JOSS?
