### 2019-09-14 #1727 @arokem
👋 @jkahn : would you be willing to review this submission for JOSS?
