+++ fatf/transparency/sklearn/surrogate_explainers.py
@@ -10,6 +10,11 @@ import warnings
 
 import numpy as np
 
+import sklearn.linear_model
+import sklearn.tree
+
+import scipy.stats
+
 import fatf.utils.data.feature_selection.sklearn as fudfs
 import fatf.transparency.sklearn.linear_model as ftslm
 import fatf.transparency.sklearn.tools as ftst
@@ -17,12 +22,11 @@ import fatf.transparency.sklearn.tools as ftst
 import fatf.utils.data.augmentation as fuda
 import fatf.utils.data.discretisation as fudd
 import fatf.utils.distances as fud
-import fatf.utils.kernels as fatf_kernels
+import fatf.utils.kernels as fuk
 import fatf.utils.transparency.explainers as fute
 import fatf.utils.array.validation as fuav
 import fatf.utils.models.validation as fumv
 import fatf.utils.array.tools as fuat
-import fatf.utils.data.transformation as fudt
 
 from fatf.exceptions import IncompatibleModelError, IncorrectShapeError
 
@@ -231,7 +235,6 @@ class SurrogateExplainer(abc.ABC):
         indices = fuat.indices_by_type(dataset)
         cat_indices = set(indices[1])
         num_indices = set(indices[0])
-        cat_indices = set(indices[1])
         all_indices = num_indices.union(cat_indices)
 
         if categorical_indices is None:
@@ -259,6 +262,8 @@ class SurrogateExplainer(abc.ABC):
             self.prediction_function = global_model.predict_proba
         else:
             self.prediction_function = global_model.predict
+
+        self.probabilistic = probabilistic
 
         if self.is_structured:
             row = np.array([dataset[0]], dtype=self.dataset.dtype)
@@ -385,16 +390,18 @@ class SurrogateExplainer(abc.ABC):
                                   'overwritten.')
 
 
-class TabularLime(SurrogateExplainer):
+class TabularLIME(SurrogateExplainer):
     """
-    An implemented of LIME explainer.
+    An implementation of LIME explainer.
 
-    Normal sampling around instance if ``sample_around_instance`` is
-    ``True`` else samples around the mean of the dataset. Then computes
-    distances between sampled points and ``data_row`` and uses K-LASSO to
-    select ``features_number`` number of features to use in the explanation.
+    Samples from the discretise dataset so the sample data has the same
+    proportions of bins as the original dataset. Then computes distances
+    between sampled points and ``data_row`` and uses K-LASSO to select
+    ``features_number`` number of features to use in the explanation.
     Then fits a local ridge regression model. The weightings of this ridge
     regression is then interpreted as feature importances.
+
+    Indicies in the parameter ``categorical_indices`` will not be discretised.
 
     For additional parameters, warnings and errors please see the parent class
     :class:`fatf.transparency.sklearn.surrogate_explainers.SurrogateExplainer`.
@@ -407,7 +414,19 @@ class TabularLime(SurrogateExplainer):
 
     Attributes
     ----------
-
+    discretiser : fatf.utils.data.discretisation.QuartileDiscretiser
+        Quartile discretiser to be used to discretise dataset and ``data_row``
+        when ``explain_instance`` is called.
+    augmentor : fatf.utils.data.augmentation.NormalSampling
+        Augmentor for performing normal sampling on the data.
+    discretised_dataset : numpy.ndarray
+        A discretised version of the dataset.
+    bin_sampling_values : Dictionary[Index, Dictionary[value, Tuple(float, float, float, float)]]
+        A dictionary specifying for each feature, a dictionary that maps the
+        discretised values in the dataset to the (minimum, maximum, mean,
+        standard deviation) for the original data bin corresponding to the
+        value. This is used to go from discretised sampled values to the
[…]
