+++ fatf/transparency/sklearn/tests/test_surrogate_explainers.py
@@ -4,7 +4,7 @@ Tests surrogate_explainers classes.
 # Author: Alex Hepburn <ah13558@bristol.ac.uk>
 # License: new BSD
 
-from typing import Dict, Union
+from typing import Dict, Union, Tuple
 
 import pytest
 
@@ -16,16 +16,14 @@ import fatf
 
 import fatf.transparency.sklearn.surrogate_explainers as ftsse
 import fatf.utils.models as fum
+import fatf.utils.data.datasets as fatf_datasets
 import fatf.utils.data.augmentation as fuda
 import fatf.utils.data.discretisation as fudd
-import fatf.utils.transparency.explainers as fute
-import fatf.utils.distances as fud
-import fatf.utils.kernels as fuk
-import fatf.transparency.sklearn.linear_model as ftslm
 
 from fatf.exceptions import IncompatibleModelError, IncorrectShapeError
 
 Index = Union[int, str]
+
 
 ONE_D_ARRAY = np.array([0, 4, 3, 0])
 NUMERICAL_NP_ARRAY_TARGET = np.array([2, 0, 1, 1, 0, 2])
@@ -50,6 +48,9 @@ MIXED_ARRAY = np.array([(0, 'a', 0.08, 'a'), (0, 'f', 0.03, 'bb'),
                         (0, 'c', 0.36, 'b'), (1, 'f', 0.07, 'bb')],
                        dtype=[('a', 'i'), ('b', 'U1'), ('c', 'f'), ('d',
                                                                     'U2')])
+
+# Test LIME and bLIMEy on IRIS dataset
+IRIS_DATASET = fatf_datasets.load_iris()
 
 
 class InvalidModel(object):
@@ -112,6 +113,31 @@ def _is_explanation_equal(dict1: Dict[str, Dict[Index, np.float64]],
                     feat_vals2 = class_exp2[feature]
                     if not np.isclose(feat_vals1, feat_vals2, tol):
                         equal = False
+            else:
+                equal = False
+    else:
+        equal = False
+    return equal
+
+def _is_bin_sampling_equal(dict1: Dict[str, Dict[Index, Tuple[float, ...]]],
+                           dict2: Dict[str, Dict[Index, Tuple[float, ...]]],
+                           tol: float = 1e-2) -> bool:
+    """
+    Tests if two bin sampling dictionarys are equal within a tolerance.
+    """
+    equal = True
+
+    if set(dict1.keys()) == set(dict2.keys()):
+        for key in dict1.keys():
+            class_exp1 = dict1[key]
+            class_exp2 = dict2[key]
+            if set(class_exp1.keys()) == set(class_exp2.keys()):
+                for feature in class_exp1.keys():
+                    feat_vals1 = class_exp1[feature]
+                    feat_vals2 = class_exp2[feature]
+                    for (val1, val2) in zip(feat_vals1, feat_vals2):
+                        if not np.isclose(val1, val2, tol):
+                            equal = False
             else:
                 equal = False
     else:
@@ -486,3 +512,389 @@ class TestSurrogateExplainer():
             self.categorical_np_dummy_surrogate.explain_instance(
                 np.array(['a', 'b']))
         assert str(exin.value) == incorrect_shape_features
+
+
+class TestTabularBlimeyTree():
+    """
+    Tests :class:`fatf.transparency.sklearn.surrogate_explainers.\
+    TabularBlimeyTree` abstract class.
+    """
+    numerical_np_array_classifier = fum.KNN(k=3)
+    numerical_np_array_classifier.fit(NUMERICAL_NP_ARRAY,
+                                      NUMERICAL_NP_ARRAY_TARGET)
+
[…]
