### 2019-09-17 #1727 @So-Cool
Hi, from the provided list @sbrugman, @giodegas and @soroushj could be potential reviewers. More generally, anybody with a machine learning background and interest in algorithmic fairness, accountability and transparency will be a good fit.
