### 2019-09-18 db7ba9a
Finished writing tests for TabularLIME
