+++ fatf/transparency/sklearn/tests/test_surrogate_explainers.py
@@ -534,11 +534,11 @@ class TestTabularBlimeyTree():
     iris_classifier = fum.KNN(k=3)
     iris_classifier.fit(IRIS_DATASET['data'], IRIS_DATASET['target'])
 
-    numerical_np_tabular_lime = ftsse.TabularBlimeyTree(
+    numerical_np_tabular_blimey = ftsse.TabularBlimeyTree(
         NUMERICAL_NP_ARRAY,
         numerical_np_array_classifier)
 
-    numerical_np_cat_tabular_lime= ftsse.TabularBlimeyTree(
+    numerical_np_cat_tabular_blimey= ftsse.TabularBlimeyTree(
         NUMERICAL_NP_ARRAY,
         numerical_np_array_classifier,
         categorical_indices=[0, 1])
@@ -631,68 +631,70 @@ class TestTabularBlimeyTree():
 
         numerical_np_explanation = {
             'class 0': {
-                'feature 0': 0.074,
-                'feature 1': 0.797,
-                'feature 2': 0.130,
+                'feature 0': 0.552,
+                'feature 1': 0.448,
+                'feature 2': 0.0,
                 'feature 3': 0.0},
             'class 1': {
                 'feature 0': 0.0,
-                'feature 1': 0.149,
-                'feature 2': 0.851,
+                'feature 1': 1.0,
+                'feature 2': 0.0,
                 'feature 3': 0.0},
             'class 2': {
-                'feature 0': 0.052,
-                'feature 1': 0.408,
-                'feature 2': 0.540,
+                'feature 0': 0.564,
+                'feature 1': 0.436,
+                'feature 2': 0.0,
                 'feature 3': 0.0}}
 
         numerical_np_cat_explanation = {
             'class 0': {
-                'feature 0': 0.071,
-                'feature 1': 0.460,
-                'feature 2': 0.197,
-                'feature 3': 0.272},
+                'feature 0': 0.197,
+                'feature 1': 0.242,
+                'feature 2': 0.267,
+                'feature 3': 0.295},
             'class 1': {
                 'feature 0': 0.0,
-                'feature 1': 0.534,
-                'feature 2': 0.466,
-                'feature 3': 0.0},
+                'feature 1': 0.086,
+                'feature 2': 0.847,
+                'feature 3': 0.067},
             'class 2': {
-                'feature 0': 0.166,
-                'feature 1': 0.151,
+                'feature 0': 0.083,
+                'feature 1': 0.190,
                 'feature 2': 0.0,
-                'feature 3': 0.683}}
+                'feature 3': 0.727}}
 
         iris_explanation = {
             'setosa': {
-                'petal length (cm)': 0.0044,
-                'petal width (cm)': 0.996,
+                'petal length (cm)': 0.991,
+                'petal width (cm)': 0.0,
                 'sepal length (cm)': 0.0,
-                'sepal width (cm)': 0.0},
+                'sepal width (cm)': 0.008},
             'versicolor': {
-                'petal length (cm)': 0.0,
-                'petal width (cm)': 1.0,
+                'petal length (cm)': 0.789,
+                'petal width (cm)': 0.190,
                 'sepal length (cm)': 0.0,
-                'sepal width (cm)': 0.0},
+                'sepal width (cm)': 0.019},
             'virginica': {
-                'petal length (cm)': 0.0,
-                'petal width (cm)': 0.093,
-                'sepal length (cm)': 0.817,
-                'sepal width (cm)': 0.089}}
+                'petal length (cm)': 0.138,
+               'petal width (cm)': 0.789,
+               'sepal length (cm)': 0.036,
+               'sepal width (cm)': 0.035}}
 
         explanation = self.numerical_np_tabular_blimey.explain_instance(
[…]
