### 2019-09-19 #1727 @arokem
@sbrugman : 👋 would you be willing to review this article for JOSS?
