+++ fatf/utils/transparency/tests/test_model_comparison.py
@@ -1,229 +1,0 @@
-"""
-Tests model_comparison functions.
-"""
-# Author: Alex Hepburn <ah13558@bristol.ac.uk>
-# License: new BSD
-
-from typing import Dict, Union
-
-import pytest
-
-import numpy as np
-
-import fatf
-import fatf.utils.data.augmentation as fuda
-import fatf.utils.models.models as fumm
-import fatf.utils.metrics.metrics as fummet
-import fatf.utils.metrics.tools as fumt
-import fatf.utils.transparency.model_comparison as futmc
-
-from fatf.exceptions import IncompatibleModelError, IncorrectShapeError
-
-NUMERICAL_NP_ARRAY = np.array([[0, 0, 0.08, 0.69], [1, 0, 0.03, 0.29],
-                               [0, 1, 0.99, 0.82], [2, 1, 0.73, 0.48],
-                               [1, 0, 0.36, 0.89], [0, 1, 0.07, 0.21]])
-NUMERICAL_STRUCT_ARRAY = np.array([(0, 0, 0.08, 0.69), (1, 0, 0.03, 0.29),
-                                   (0, 1, 0.99, 0.82), (2, 1, 0.73, 0.48),
-                                   (1, 0, 0.36, 0.89), (0, 1, 0.07, 0.21)],
-                                  dtype=[('a', 'i'), ('b', 'i'), ('c', 'f'),
-                                         ('d', 'f')])
-NUMERICAL_NP_ARRAY_LOCAL = np.array([[0, 0, 0.09, 0.70], [1, 0, 0.06, 0.60],
-                                     [0, 0, 0.05, 0.50], [0, 1, 0.99, 0.82]])
-NUMERICAL_STRUCT_ARRAY_LOCAL = np.array([(0, 0, 0.09, 0.70), (1, 0, 0.06, 0.60),
-                                         (0, 0, 0.05, 0.50), (0, 1, 0.99, 0.82)],
-                                         dtype=[('a', 'i'), ('b', 'i'), ('c', 'f8'),
-                                                ('d', 'f8')])
-
-NUMERICAL_NP_ARRAY_TARGET = np.array([2, 0, 1, 1, 0, 2])
-NUMERICAL_NP_ARRAY_LOCAL_TARGET = np.array([1, 0, 1, 1])
-
-
-def test_input_is_valid():
-    """
-    Tests :func:`fatf.utils.transparency.model_comparison._input_is_valid`.
-    """
-    incorrect_shape_dataset = ('The input dataset must be a 2-dimensional '
-                               'numpy array.')
-    type_error_dataset = 'The input dataset must be of a base type.'
-    incorrect_shape_datarow = ('The data_row must either be a 1-dimensional '
-                               'numpy array or numpy void object for '
-                               'structured rows.')
-    incorrect_dtype_data = ('The dtype of the data_row is different to '
-                            'the dtype of the dataset array.')
-    features_index_error = ('The following indices are invalid for the input '
-                            'dataset: {}.')
-    features_type_error = ('The local_features parameter must be a Python '
-                           'list or None.')
-    datarow_features_error = ('The data_row must contain the same number of '
-                              'features as the dataset.')
-    global_model_error = ('This functionality requires the global model to '
-                          'be capable of outputting predicted class '
-                          'probabilities via predict proba method.')
-    local_model_error = ('This functionality requires the local model to be '
-                         'capable of outputting predicted class via '
-                         'predict method.')
-    r_fid_type_err = 'r_fid must be a float.'
[…]
