+++ fatf/utils/transparency/tests/test_surrogate_evaluation.py
@@ -1,0 +1,390 @@
+"""
+Tests surrogate evaluation metrics.
+"""
+# Author: Alex Hepburn <ah13558@bristol.ac.uk>
+#         Kacper Sokol <k.sokol@bristol.ac.uk>
+# License: new BSD
+
+from typing import Dict, Union
+
+import pytest
+
+import numpy as np
+
+from fatf.exceptions import IncompatibleModelError, IncorrectShapeError
+
+import fatf
+import fatf.utils.data.augmentation as fuda
+import fatf.utils.metrics.metrics as fummet
+import fatf.utils.metrics.tools as fumt
+import fatf.utils.models.models as fumm
+import fatf.utils.transparency.surrogate_evaluation as futs
+
+NUMERICAL_NP_ARRAY = np.array([[0, 0, 0.08, 0.69], [1, 0, 0.03, 0.29],
+                               [0, 1, 0.99, 0.82], [2, 1, 0.73, 0.48],
+                               [1, 0, 0.36, 0.89], [0, 1, 0.07, 0.21]])
+NUMERICAL_STRUCT_ARRAY = np.array([(0, 0, 0.08, 0.69), (1, 0, 0.03, 0.29),
+                                   (0, 1, 0.99, 0.82), (2, 1, 0.73, 0.48),
+                                   (1, 0, 0.36, 0.89), (0, 1, 0.07, 0.21)],
+                                  dtype=[('a', 'i'), ('b', 'i'), ('c', 'f'),
+                                         ('d', 'f')])
+NUMERICAL_NP_ARRAY_LOCAL = np.array([[0, 0, 0.09, 0.70], [1, 0, 0.06, 0.60],
+                                     [0, 0, 0.05, 0.50], [0, 1, 0.99, 0.82]])
+NUMERICAL_STRUCT_ARRAY_LOCAL = np.array(
+    [(0, 0, 0.09, 0.7), (1, 0, 0.06, 0.60), (0, 0, 0.05, 0.5),
+     (0, 1, 0.99, 0.82)],
+    dtype=[('a', 'i'), ('b', 'i'), ('c', 'f8'), ('d', 'f8')])
+
+NUMERICAL_NP_ARRAY_TARGET = np.array([2, 0, 1, 1, 0, 2])
+NUMERICAL_NP_ARRAY_LOCAL_TARGET = np.array([1, 0, 1, 1])
+
+
+def test_validate_input_local_fidelity():
+    """
+    Tests the ``_validate_input_local_fidelity`` function.
+
+    This function tests the :func:`fatf.utils.transparency.\
+surrogate_evaluation._validate_input_local_fidelity` function.
+    """
+    incorrect_shape_dataset = ('The input dataset must be a 2-dimensional '
+                               'numpy array.')
+    type_error_dataset = ('The input dataset must be of a base type -- '
+                          'numbers and/or strings.')
+    incorrect_shape_datarow = ('The data_row must either be a 1-dimensional '
+                               'numpy array or a numpy void object for '
+                               'structured data rows.')
+    incorrect_dtype_data = ('The dtype of the data_row is too different from '
+                            'the dtype of the dataset array.')
+    datarow_features_error = ('The data_row must contain the same number of '
+                              'features as the dataset.')
+
+    global_model_incompatible = ('The global predictive function must have '
+                                 'exactly *one* required parameter to work '
+                                 'with this metric.')
+    global_model_type = ('The global_predictive_function should be a Python '
+                         'callable, e.g., a Python function.')
+    local_model_incompatible = ('The local predictive function must have '
+                                'exactly *one* required parameter to work '
+                                'with this metric.')
+    local_model_type = ('The local_predictive_function should be a Python '
[…]

[…]
