+++ fatf/utils/data/tests/test_instance_sampler.py
@@ -1,146 +1,0 @@
-"""
-Tests instance samplers.
-"""
-# Author: Alex Hepburn <ah13558@bristol.ac.uk>
-#         Kacper Sokol <k.sokol@bristol.ac.uk>
-# License: new BSD
-
-import numpy as np
-
-import pytest
-
-import fatf
-import fatf.utils.data.instance_sampler as fuis
-from fatf.exceptions import IncorrectShapeError
-
-# yapf: disable
-NUMERICAL_NP_ARRAY = np.array([
-    [0, 0, 0.08, 0.69],
-    [1, 0, 0.03, 0.29],
-    [0, 1, 0.99, 0.82],
-    [2, 1, 0.73, 0.48],
-    [1, 0, 0.36, 0.89],
-    [0, 1, 0.07, 0.21]])
-NUMERICAL_STRUCT_ARRAY = np.array(
-    [(0, 0, 0.08, 0.69),
-     (1, 0, 0.03, 0.29),
-     (0, 1, 0.99, 0.82),
-     (2, 1, 0.73, 0.48),
-     (1, 0, 0.36, 0.89),
-     (0, 1, 0.07, 0.21)],
-    dtype=[('a', 'i'), ('b', 'i'), ('c', 'f'), ('d', 'f')])
-CATEGORICAL_NP_ARRAY = np.array([
-    ['a', 'b', 'c'],
-    ['a', 'f', 'g'],
-    ['b', 'c', 'c'],
-    ['b', 'f', 'c'],
-    ['a', 'f', 'c'],
-    ['a', 'b', 'g']])
-CATEGORICAL_STRUCT_ARRAY = np.array(
-    [('a', 'b', 'c'),
-     ('a', 'f', 'g'),
-     ('b', 'c', 'c'),
-     ('b', 'f', 'c'),
-     ('a', 'f', 'c'),
-     ('a', 'b', 'g')],
-    dtype=[('a', 'U1'), ('b', 'U1'), ('c', 'U1')])
-MIXED_ARRAY = np.array(
-    [(0, 'a', 0.08, 'a'),
-     (0, 'f', 0.03, 'bb'),
-     (1, 'c', 0.99, 'aa'),
-     (1, 'a', 0.73, 'a'),
-     (0, 'c', 0.36, 'b'),
-     (1, 'f', 0.07, 'bb')],
-    dtype=[('a', 'i'), ('b', 'U1'), ('c', 'f'), ('d', 'U2')])
-# yapf: enable
-
-
-def test_validate_input():
-    """
-    Tests :func:`fatf.utils.data.instance_sampler._validate_input`.
-    """
-    shape_msg = ('The data_row must either be a 1-dimensional numpy array or '
-                 'numpy void object for structured rows.')
-    samples_msg  = 'The samples_number parameter must be a positive integer.'
-    samples_type_msg = 'The samples_number parameter must be an integer.'
-
-    with pytest.raises(IncorrectShapeError) as exin:
-        fuis._validate_input(NUMERICAL_NP_ARRAY, 50)
-    assert str(exin.value) == shape_msg
-
-    with pytest.raises(ValueError) as exin:
-        fuis._validate_input(NUMERICAL_NP_ARRAY[0], -1)
-    assert str(exin.value) == samples_msg
-
-    with pytest.raises(TypeError) as exin:
-        fuis._validate_input(NUMERICAL_NP_ARRAY[0], 'a')
-    assert str(exin.value) == samples_type_msg
-
-    # Everything good
-    assert fuis._validate_input(NUMERICAL_NP_ARRAY[0], 50)
-    assert fuis._validate_input(CATEGORICAL_NP_ARRAY[0], 50)
[…]
