### 2019-09-24 #1727 @arokem
@giodegas or @soroushj: would either of you be able to review this article for JOSS?
