### 2019-09-25 #1727 @arokem
👋 @zfurnas : would you be willing to review this article for JOSS?
