+++ fatf/transparency/sklearn/tests/test_blimey.py
@@ -1,835 +1,0 @@
-"""
-Tests Blimey class.
-"""
-# Author: Alex Hepburn <ah13558@bristol.ac.uk>
-# License: new BSD
-
-from typing import Dict, Union
-
-import pytest
-
-import numpy as np
-
-from sklearn.linear_model import Ridge
-
-import fatf
-
-import fatf.transparency.sklearn.blimey as ftsmb
-import fatf.utils.models as fum
-import fatf.utils.data.augmentation as fuda
-import fatf.utils.data.discretisation as fudd
-import fatf.utils.transparency.explainers as fute
-import fatf.utils.distances as fud
-import fatf.utils.kernels as fuk
-import fatf.transparency.sklearn.linear_model as ftslm
-
-from fatf.exceptions import IncompatibleModelError, IncorrectShapeError
-
-Index = Union[int, str]
-
-ONE_D_ARRAY = np.array([0, 4, 3, 0])
-NUMERICAL_NP_ARRAY_TARGET = np.array([2, 0, 1, 1, 0, 2])
-NUMERICAL_NP_ARRAY = np.array([[0, 0, 0.08, 0.69], [1, 0, 0.03, 0.29],
-                               [0, 1, 0.99, 0.82], [2, 1, 0.73, 0.48],
-                               [1, 0, 0.36, 0.89], [0, 1, 0.07, 0.21]])
-NUMERICAL_STRUCT_ARRAY = np.array([(0, 0, 0.08, 0.69), (1, 0, 0.03, 0.29),
-                                   (0, 1, 0.99, 0.82), (2, 1, 0.73, 0.48),
-                                   (1, 0, 0.36, 0.89), (0, 1, 0.07, 0.21)],
-                                  dtype=[('a', 'i'), ('b', 'i'), ('c', 'f'),
-                                         ('d', 'f')])
-CATEGORICAL_NP_ARRAY = np.array([['a', 'b', 'c'], ['a', 'f', 'g'],
-                                 ['b', 'c', 'c'], ['b', 'f', 'c'],
-                                 ['a', 'f', 'c'], ['a', 'b', 'g']])
-CATEGORICAL_STRUCT_ARRAY = np.array([('a', 'b', 'c'), ('a', 'f', 'g'),
-                                     ('b', 'c', 'c'), ('b', 'f', 'c'),
-                                     ('a', 'f', 'c'), ('a', 'b', 'g')],
-                                    dtype=[('a', 'U1'), ('b', 'U1'), ('c',
-                                                                      'U1')])
-MIXED_ARRAY = np.array([(0, 'a', 0.08, 'a'), (0, 'f', 0.03, 'bb'),
-                        (1, 'c', 0.99, 'aa'), (1, 'a', 0.73, 'a'),
-                        (0, 'c', 0.36, 'b'), (1, 'f', 0.07, 'bb')],
-                       dtype=[('a', 'i'), ('b', 'U1'), ('c', 'f'), ('d',
-                                                                    'U2')])
-
-NUMERICAL_NP_BLIMEY = {
-    'Class A': {
-        '0.07 < *C* <= 0.22': -0.0152,
-        '0.58 < *D* <= 0.79': 0.0821,
-        '*A* <= 0.00': -0.3665,
-        '*B* <= 0.00': 0.1791,
-    },
-    'Class B': {
-        '0.07 < *C* <= 0.22': 0.0546,
-        '0.58 < *D* <= 0.79': 0.0073,
-        '*A* <= 0.00': 0.1489,
-        '*B* <= 0.00': -0.0791
-    },
-    'Class C': {
[…]
