+++ fatf/transparency/sklearn/blimey.py
@@ -1,624 +1,0 @@
-"""
-Blimey implementation.
-"""
-# Author: Alex Hepburn <ah13558@bristol.ac.uk>
-# License: new BSD
-
-from typing import Dict, Union, Optional, List, Callable
-import warnings
-
-import numpy as np
-
-import fatf.utils.data.feature_selection.sklearn as fudfs
-import fatf.transparency.sklearn.linear_model as ftslm
-import fatf.transparency.sklearn.tools as ftst
-
-import fatf.utils.data.augmentation as fuda
-import fatf.utils.data.discretisation as fudd
-import fatf.utils.distances as fud
-import fatf.utils.kernels as fatf_kernels
-import fatf.utils.transparency.explainers as fute
-import fatf.utils.array.validation as fuav
-import fatf.utils.models.validation as fumv
-import fatf.utils.array.tools as fuat
-import fatf.utils.data.transformation as fudt
-
-from fatf.exceptions import IncompatibleModelError, IncorrectShapeError
-
-__all__ = ['Blimey']
-
-Index = Union[int, str]
-
-
-def _input_is_valid(dataset: np.ndarray,
-                    augmentor: fuda.Augmentation,
-                    explainer: fute.Explainer,
-                    global_model: object,
-                    local_model: object,
-                    categorical_indices: Optional[List[Index]] = None,
-                    class_names: Optional[List[str]] = None,
-                    feature_names: Optional[List[str]] = None,
-                    discretise_first: Optional[bool] = None,
-                    discretiser: Optional[fudd.Discretiser] = None) -> bool:
-    """
-    Validates the input parameters of Blimey class.
-
-    For the input parameter description, warnings and exceptions please see
-    the documentation of the :func`fatf.transparency.models.blimey.Blimey.
-    __init__` function.
-
-    Returns
-    -------
-    is_input_ok : boolean
-        ``True`` if input is valid, ``False`` otherwise.
-    """
-    is_input_ok = False
-
-    if not fuav.is_2d_array(dataset):
-        raise IncorrectShapeError('The input dataset must be a 2-dimensional '
-                                  'array.')
-
-    if not fuav.is_base_array(dataset):
-        raise TypeError('The input dataset must only contain base types '
-                        '(textual and numerical).')
-
-    if not fumv.check_model_functionality(
-            global_model, require_probabilities=True, suppress_warning=True):
-        raise IncompatibleModelError(
-            'This functionality requires the global model to be capable of '
-            'outputting probabilities via predict_proba method.')
-
-    if not fumv.check_model_functionality(
-            local_model,
-            require_probabilities=False,
-            suppress_warning=True):
-        raise IncompatibleModelError(
-            'This functionality requires the local model to be capable of '
-            'outputting predictions via predict method.')
-
-    if not issubclass(explainer, fute.Explainer):
-        raise TypeError('The explainer object must inherit from abstract '
-                        'class fatf.utils.transparency.explainers.Explainer.')
-
-    if categorical_indices is not None:
-        if isinstance(categorical_indices, list):
-            invalid_indices = fuat.get_invalid_indices(
-                dataset, np.asarray(categorical_indices))
-            if invalid_indices.size:
-                raise IndexError('The following indices are invalid for the '
-                                 'input dataset: {}.'.format(invalid_indices))
-        else:
-            raise TypeError('The categorical_indices parameter must be a '
-                            'Python list or None.')
-
-    if not issubclass(augmentor, fuda.Augmentation):
-        raise TypeError('The augmentor object must inherit from abstract '
-                        'class fatf.utils.augmentation.Augmentation.')
-
-    if (discretiser is not None
-            and not issubclass(discretiser, fudd.Discretiser)):
[…]
