+++ fatf/utils/data/tests/test_augmentation.py
@@ -1,5 +1,5 @@
 """
-Functions for testing data augmentation classes.
+Functions and classes for testing data augmentation approaches.
 """
 # Author: Alex Hepburn <ah13558@bristol.ac.uk>
 #         Kacper Sokol <k.sokol@bristol.ac.uk>
@@ -1033,50 +1033,68 @@ class TestMixup(object):
 
 def get_truncated_mean_std(minimum, maximum, original_mean, original_std):
         """
-        Function for getting the theoretical truncated normal mean and std
-        from the original normal mean and std, and the minimum and maximum
-        for which values are truncated.
+        Computes the theoretical mean and standard deviation of a truncated
+        normal distribution from its initialisation parameters: the original
+        normal mean and standard deviation, and the minimum and maximum within
+        which values are truncated.
 
-        Equations for calculating these can be found at:
-        https://en.wikipedia.org/wiki/Truncated_normal_distribution
+        Equations for calculating these -- implemented by this function -- can
+        be found here_.
+
+        .. _here: https://en.wikipedia.org/wiki/Truncated_normal_distribution
         """
         def cdf(epsilon):
-            return 1/2 * (1 + scipy.special.erf(epsilon/np.sqrt(2)))
+            return (1 / 2) * (1 + scipy.special.erf(epsilon / np.sqrt(2)))
 
         def norm(episilon):
-            return (1/np.sqrt(2*np.pi) * np.exp(-1/2 * episilon**2))
+            return 1 / np.sqrt(2 * np.pi) * np.exp(-1 / 2 * episilon**2)
 
-        alpha = (minimum-original_mean) / original_std
-        beta = (maximum-original_mean) / original_std
+        alpha = (minimum - original_mean) / original_std
+        beta = (maximum - original_mean) / original_std
         z_phi = cdf(beta) - cdf(alpha)
-        mean = original_mean + ((norm(alpha) - norm(beta)) / z_phi) * \
-               original_std
-        std = np.sqrt(original_std ** 2 * (1 + (alpha * norm(alpha) - beta * \
-              norm(beta)) / z_phi - ((norm(alpha) - norm(beta)) / z_phi) ** 2 ))
-        return mean, std
+
+        n_ab = norm(alpha) - norm(beta)
+
+        computed_mean = (
+            original_mean + (n_ab / z_phi) * original_std)
+        computed_var = (
+            original_std**2 *
+            (1 + (alpha * norm(alpha) - beta * norm(beta)) / z_phi
+             - (n_ab / z_phi)**2)
+        )
+        computed_std = np.sqrt(computed_var)
+
+        return computed_mean, computed_std
 
 
 class TestTruncatedNormalSampling(object):
     """
     Tests :class:`fatf.utils.data.augmentation.TruncatedNormalSampling` class.
     """
-    numerical_np_0_augmentor = fuda.TruncatedNormalSampling(NUMERICAL_NP_ARRAY, [0])
     numerical_np_augmentor = fuda.TruncatedNormalSampling(NUMERICAL_NP_ARRAY)
-    numerical_struct_a_augmentor = fuda.TruncatedNormalSampling(NUMERICAL_STRUCT_ARRAY,
-                                                       ['a'])
-    numerical_struct_augmentor = fuda.TruncatedNormalSampling(NUMERICAL_STRUCT_ARRAY)
+    numerical_np_0_augmentor = fuda.TruncatedNormalSampling(
+        NUMERICAL_NP_ARRAY, [0])
+
+    numerical_struct_augmentor = fuda.TruncatedNormalSampling(
+        NUMERICAL_STRUCT_ARRAY)
[…]
