+++ fatf/utils/data/augmentation.py
@@ -604,7 +604,7 @@ Tuple[numpy.ndarray, numpy.ndarray]]
 
         For the documentation of parameters, warnings and errors please see the
         description of the
-        :func:`~fatf.utils.data.augmentation.Augmentation.sample` method in the
+        :func:`fatf.utils.data.augmentation.Augmentation.sample` method in the
         parent :class:`fatf.utils.data.augmentation.Augmentation` class.
         """
         assert self._validate_sample_input(data_row,
@@ -1230,7 +1230,7 @@ def _validate_input_normalclassdiscovery(
                              'be a number between 0 and 1 (not inclusive).')
     else:
         raise TypeError('The class_proportion_threshold parameter is not a '
-                        'float.')
+                        'number.')
 
     if isinstance(standard_deviation_init, Number):
         if standard_deviation_init <= 0:
@@ -1238,7 +1238,7 @@ def _validate_input_normalclassdiscovery(
                              'positive number (greater than 0).')
     else:
         raise TypeError('The standard_deviation_init parameter is not a '
-                        'float.')
+                        'number.')
 
     if isinstance(standard_deviation_increment, Number):
         if standard_deviation_increment <= 0:
@@ -1246,7 +1246,7 @@ def _validate_input_normalclassdiscovery(
                              'be a positive number (greater than 0).')
     else:
         raise TypeError('The standard_deviation_increment parameter is not a '
-                        'float.')
+                        'number.')
 
     is_valid = True
     return is_valid
@@ -1272,7 +1272,7 @@ class NormalClassDiscovery(Augmentation):
     probability for each unique value calculated based on the frequency of
     their appearance in the dataset.
 
-    .. note:: The number of classes for classifiers.
+    .. note:: The number of classes when using a *classifier*.
 
        Consider using the ``classes_number`` parameter when using a
        non-probabilistic ``predictive_function``. For more details please see
@@ -1320,12 +1320,25 @@ class NormalClassDiscovery(Augmentation):
         The minimum proportion of data points assigned to a different class
         by the ``predictive_function`` when sampling for each data point as per
         the procedure described above.
-    standard_deviation_init : float, optional (default=0.01)
+
+        .. warning:: Setting the ``class_proportion_threshold`` parameter.
+
+           This augmenter samples a cloud of points for each discovered class
+           with each cloud having 1 / ``classes_number`` number of points.
+           This means that the value of the ``class_proportion_threshold`` has
+           to be smaller than this number for the sampling to be successful.
+           For example, for 2 classes and 100 sampled points, 2 clouds of 50
+           data points each will be generated. By setting the
+           ``class_proportion_threshold`` parameter to ``0.6``, at least 60
+           point of each class are expected, which cannot be achieved.
+
+    standard_deviation_init : float, optional (default=1)
         The standard deviation of the normal distribution used for initial
         sampling around each selected data point.
-    standard_deviation_increment : float, optional (default=0.001)
+    standard_deviation_increment : float, optional (default=0.1)
         The increment used to increase the standard deviation every time the
-        sample does not satisfy the specified ``class_proportion_threshold``.
+        sample does not satisfy the specified ``class_proportion_threshold``
+        or at least one data point of yet unseen class is not discovered.
 
     Raises
     ------
@@ -1335,11 +1348,16 @@ class NormalClassDiscovery(Augmentation):
     RuntimeError
         The class initialisation was unable to identify the number of classes
         using the input ``dataset`` and the provided ``predictive_function``.
+        The value of the ``class_proportion_threshold`` parameter is too large
+        for the given number of classes (please see the warning in the
+        ``class_proportion_threshold`` parameter description for more
+        information).
     TypeError
         The ``predictive_function`` is not a Python callable. The
[…]
