### 2019-09-30 #1727 @arokem
👋 @ahurriyetoglu : would you be willing to review this article for JOSS?
