### 2019-10-02 063ade1
Linting and typing
