+++ fatf/utils/data/tests/test_augmentation.py
@@ -1032,39 +1032,39 @@ class TestMixup(object):
 
 
 def get_truncated_mean_std(minimum, maximum, original_mean, original_std):
-        """
-        Computes the theoretical mean and standard deviation of a truncated
-        normal distribution from its initialisation parameters: the original
-        normal mean and standard deviation, and the minimum and maximum within
-        which values are truncated.
+    """
+    Computes the theoretical mean and standard deviation of a truncated
+    normal distribution from its initialisation parameters: the original
+    normal mean and standard deviation, and the minimum and maximum within
+    which values are truncated.
 
-        Equations for calculating these -- implemented by this function -- can
-        be found here_.
+    Equations for calculating these -- implemented by this function -- can
+    be found here_.
 
-        .. _here: https://en.wikipedia.org/wiki/Truncated_normal_distribution
-        """
-        def cdf(epsilon):
-            return (1 / 2) * (1 + scipy.special.erf(epsilon / np.sqrt(2)))
+    .. _here: https://en.wikipedia.org/wiki/Truncated_normal_distribution
+    """
+    def cdf(epsilon):
+        return (1 / 2) * (1 + scipy.special.erf(epsilon / np.sqrt(2)))
 
-        def norm(episilon):
-            return 1 / np.sqrt(2 * np.pi) * np.exp(-1 / 2 * episilon**2)
+    def norm(episilon):
+        return 1 / np.sqrt(2 * np.pi) * np.exp(-1 / 2 * episilon**2)
 
-        alpha = (minimum - original_mean) / original_std
-        beta = (maximum - original_mean) / original_std
-        z_phi = cdf(beta) - cdf(alpha)
+    alpha = (minimum - original_mean) / original_std
+    beta = (maximum - original_mean) / original_std
+    z_phi = cdf(beta) - cdf(alpha)
 
-        n_ab = norm(alpha) - norm(beta)
+    n_ab = norm(alpha) - norm(beta)
 
-        computed_mean = (
-            original_mean + (n_ab / z_phi) * original_std)
-        computed_var = (
-            original_std**2 *
-            (1 + (alpha * norm(alpha) - beta * norm(beta)) / z_phi
-             - (n_ab / z_phi)**2)
-        )
-        computed_std = np.sqrt(computed_var)
+    computed_mean = (
+        original_mean + (n_ab / z_phi) * original_std)
+    computed_var = (
+        original_std**2
+        * (1 + (alpha * norm(alpha) - beta * norm(beta)) / z_phi
+           - (n_ab / z_phi)**2)
+    )
+    computed_std = np.sqrt(computed_var)
 
-        return computed_mean, computed_std
+    return computed_mean, computed_std
 
 
 class TestTruncatedNormalSampling(object):
@@ -1166,7 +1166,7 @@ class TestTruncatedNormalSampling(object):
             [(0, 0.243, 0.048, 0.697),
              (1, 0.066, 0.591, 0.842),
              (1, 0.728, 0.214, 0.418)],
-             dtype=[('a', 'i'), ('b', 'f'), ('c', 'f'), ('d', 'f')])
+            dtype=[('a', 'i'), ('b', 'f'), ('c', 'f'), ('d', 'f')])
         categorical_np_results = np.array([
             ['a', 'f', 'c'],
             ['a', 'f', 'c'],
@@ -1175,7 +1175,7 @@ class TestTruncatedNormalSampling(object):
             [('a', 'b', 'g'),
              ('a', 'f', 'c'),
              ('a', 'f', 'c')],
-             dtype=[('a', 'U1'), ('b', 'U1'), ('c', 'U1')])
[…]
