+++ fatf/utils/transparency/__init__.py
@@ -1,0 +1,5 @@
+"""
+The :mod:`fatf.utils.transparency` module holds transparency utilities.
+"""
+# Author: Kacper Sokol <k.sokol@bristol.ac.uk>
+# License: new BSD
