+++ fatf/utils/data/tests/test_discretisation.py
@@ -13,6 +13,7 @@ import fatf.utils.data.discretisation as fudd
 
 from fatf.exceptions import IncorrectShapeError
 
+# yapf: disable
 NUMERICAL_NP_ARRAY = np.array([
     [0, 0, 0.08, 0.69],
     [1, 0, 0.03, 0.29],
@@ -75,6 +76,7 @@ MIXED_DISCRETISED = np.array(
      (0, 'c', 2, 'b'),
      (2, 'f', 0, 'bb')],
     dtype=[('a', 'i1'), ('b', 'U1'), ('c', 'i1'), ('d', 'U2')])
+# yapf: enable
 
 
 def test_validate_input_discretiser():
@@ -107,13 +109,13 @@ def test_validate_input_discretiser():
     assert str(exin.value) == data_type_error
 
     with pytest.raises(TypeError) as exin:
-        fudd._validate_input_discretiser(NUMERICAL_NP_ARRAY,
-                                         categorical_indices=0)
+        fudd._validate_input_discretiser(
+            NUMERICAL_NP_ARRAY, categorical_indices=0)
     assert str(exin.value) == cidx_type_error
     #
     with pytest.raises(IndexError) as exin:
-        fudd._validate_input_discretiser(MIXED_ARRAY,
-                                         categorical_indices=['f'])
+        fudd._validate_input_discretiser(
+            MIXED_ARRAY, categorical_indices=['f'])
     assert str(exin.value) == cidx_index_error.format(['f'])
     with pytest.raises(IndexError) as exin:
         fudd._validate_input_discretiser(MIXED_ARRAY, categorical_indices=[1])
@@ -128,17 +130,17 @@ def test_validate_input_discretiser():
     assert str(exin.value) == feature_names_incorrect_shape
     #
     with pytest.raises(TypeError) as exin:
-        fudd._validate_input_discretiser(NUMERICAL_NP_ARRAY,
-                                         feature_names=[0, 1, 2, 'a'])
+        fudd._validate_input_discretiser(
+            NUMERICAL_NP_ARRAY, feature_names=[0, 1, 2, 'a'])
     assert str(exin.value) == feature_name_type_error.format(0)
     #
     with pytest.raises(TypeError) as exin:
-        fudd._validate_input_discretiser(MIXED_ARRAY,
-                                         feature_names=['a', 'b', 3, 'd'])
+        fudd._validate_input_discretiser(
+            MIXED_ARRAY, feature_names=['a', 'b', 3, 'd'])
     assert str(exin.value) == feature_name_type_error.format(3)
 
-    assert fudd._validate_input_discretiser(MIXED_ARRAY,
-                                            categorical_indices=['a', 'b'])
+    assert fudd._validate_input_discretiser(
+        MIXED_ARRAY, categorical_indices=['a', 'b'])
 
 
 class TestDiscretiser():
@@ -222,34 +224,41 @@ _validate_input_discretise` method.
         assert categorical_np_discretiser.categorical_indices == [0, 1, 2]
         assert categorical_np_discretiser.numerical_indices == []
         assert (categorical_np_discretiser.feature_names_map
-                == dict(zip([0, 1, 2], ['aa', 'bb', 'cc'])))
+                == dict(zip([0, 1, 2], ['aa', 'bb', 'cc'])))  # yapf: disable
         assert categorical_np_discretiser.feature_value_names == {}
         assert categorical_np_discretiser.feature_bin_boundaries == {}
 
         categorical_struct_discretiser = self.BaseDiscretiser(
             CATEGORICAL_STRUCT_ARRAY)
-        assert (categorical_struct_discretiser.dataset_dtype
[…]
