### 2019-10-07 #1727 @giodegas
>
>
> @giodegas or @soroushj: would either of you be able to review this article for JOSS?

I would be happy to do it, but cannot allocate time at the moment.
Sorry.
