### 2019-10-14 #1727 @arokem
@FreddieV4 : would you be willing to review this article for JOSS?
