### 2019-10-14 #1727 @arokem
@tianxzhu : would you be willing to review this article for JOSS?
