### 2019-10-23 #1727 @arokem
👋 @adebayoj : would you be interested in reviewing this article for the [Journal of Open Source Software](https://joss.theoj.org/)? It is closely related to the work that you do, so your expertise will be
 very helpful here. For more information about the journal and our process, you can refer to this webpage:  https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
