### 2019-10-28 361b5e4
CI PyPi deployment and RNG seeding
