### 2019-10-28 2b97a62
Python 3.8 testing and doctest fix
