### 2019-10-29 873ab4c
scipy and numpy wheels not available for Python 3.8
