+++ .travis.yml
@@ -20,7 +20,6 @@ before_cache:
   - rm -f $HOME/.cache/pip/log/debug.log
 
 python:
-  - '3.8'
   - '3.7'
   - '3.6'
   - '3.5'
@@ -51,14 +50,6 @@ matrix:
     - python: '3.7'
       env: FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.12.1
     - python: '3.7'
-      env: FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.10.1 FATF_TEST_MATPLOTLIB=3.0.0
-    - python: '3.8'
-      env: FATF_TEST_SCIPY=1.0.1  FATF_TEST_NUMPY=1.14.6
-    - python: '3.8'
-      env: FATF_TEST_SCIPY=0.19.1 FATF_TEST_NUMPY=1.13.3
-    - python: '3.8'
-      env: FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.12.1
-    - python: '3.8'
       env: FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.10.1 FATF_TEST_MATPLOTLIB=3.0.0
     - python: '3.8-dev'
       env: FATF_TEST_SCIPY=1.0.1  FATF_TEST_NUMPY=1.14.6
