### 2019-11-02 33e5a00
Version added note and Thales URL fix
The SSL certificate of Thales' website (https://www.thalesgroup.com/en)
could not be verified, hence was removed.
