### 2019-11-02 fb11fc0
Surrogate explainers testing
