+++ fatf/utils/testing/transparency.py
@@ -1,0 +1,211 @@
+"""
+The :mod:`fatf.utils.testing.transparency` module holds transparency testing
+assets.
+"""
+# Author: Kacper Sokol <k.sokol@bristol.ac.uk>
+# License: new BSD
+
+from typing import Dict, Iterable, List, Tuple, Union
+
+import numpy as np
+
+__all__ = ['LABELS',
+           'NUMERICAL_NP_ARRAY',
+           'NUMERICAL_STRUCT_ARRAY',
+           'CATEGORICAL_NP_ARRAY',
+           'CATEGORICAL_STRUCT_ARRAY',
+           'MIXED_ARRAY',
+           'InvalidModel',
+           'NonProbabilisticModel',
+           'is_explanation_equal_list',
+           'is_explanation_equal_dict']  # yapf: disable
+
+ExplanationList = Dict[str, List[Tuple[str, Union[Iterable[float], float]]]]
+ExplanationDict = Dict[str,
+                       Dict[Union[str, int], Union[Iterable[float], float]]]
+
+# yapf: disable
+LABELS = np.array([2, 0, 1, 1, 0, 2])
+
+NUMERICAL_NP_ARRAY = np.array([
+    [0, 0, 0.08, 0.69],
+    [1, 0, 0.03, 0.29],
+    [0, 1, 0.99, 0.82],
+    [2, 1, 0.73, 0.48],
+    [1, 0, 0.36, 0.89],
+    [0, 1, 0.07, 0.21]])
+NUMERICAL_STRUCT_ARRAY = np.array(
+    [(0, 0, 0.08, 0.69),
+     (1, 0, 0.03, 0.29),
+     (0, 1, 0.99, 0.82),
+     (2, 1, 0.73, 0.48),
+     (1, 0, 0.36, 0.89),
+     (0, 1, 0.07, 0.21)],
+    dtype=[('a', int), ('b', int), ('c', float), ('d', float)])
+
+CATEGORICAL_NP_ARRAY = np.array(
+    [['a', 'b', 'c'], ['a', 'f', 'g'], ['b', 'c', 'c'],
+     ['b', 'f', 'c'], ['a', 'f', 'c'], ['a', 'b', 'g']])
+CATEGORICAL_STRUCT_ARRAY = np.array(
+    [('a', 'b', 'c'), ('a', 'f', 'g'), ('b', 'c', 'c'),
+     ('b', 'f', 'c'), ('a', 'f', 'c'), ('a', 'b', 'g')],
+    dtype=[('a', 'U1'), ('b', 'U1'), ('c', 'U1')])
+
+MIXED_ARRAY = np.array(
+    [(0, 'a', 0.08, 'a'), (0, 'f', 0.03, 'bb'), (1, 'c', 0.99, 'aa'),
+     (1, 'a', 0.73, 'a'), (0, 'c', 0.36, 'b'), (1, 'f', 0.07, 'bb')],
+    dtype=[('a', 'i'), ('b', 'U1'), ('c', 'f'), ('d', 'U2')])
+# yapf: enable
+
+
+class InvalidModel(object):
+    """
+    An invalid model class -- it does not implement a ``predict_proba`` method.
+    """
+
+    # pylint: disable=useless-object-inheritance,too-few-public-methods
+
+    def __init__(self):
+        """
+        Invalid initialisation.
+        """
+
+    def fit(self, data, targets):
+        """
+        Invalid fit.
+        """
+
+
+class NonProbabilisticModel(InvalidModel):
+    """
+    A model that is not probabilistic -- no ``predict_proba`` function.
+    """
+
+    def __init__(self, prediction_function):
+        """
+        Non-probabilistic initialisation.
+        """
+        super().__init__()
+        self.prediction_function = prediction_function
+
[…]
