+++ fatf/transparency/tests/test_surrogate_explainers.py
@@ -1,470 +1,688 @@
 """
-Tests surrogate_explainers classes.
+Tests surrogate explainers.
 """
 # Author: Alex Hepburn <ah13558@bristol.ac.uk>
+#         Kacper Sokol <K.Sokol@bristol.ac.uk>
 # License: new BSD
 
-from typing import Dict, Union, Tuple
-
+import importlib
 import pytest
+import sys
 
 import numpy as np
 
-from sklearn.linear_model import Ridge
+from fatf.exceptions import IncompatibleModelError, IncorrectShapeError
 
 import fatf
 
-import fatf.transparency.sklearn.surrogate_explainers as ftsse
-import fatf.utils.models as fum
-import fatf.utils.data.datasets as fatf_datasets
+import fatf.transparency.surrogate_explainers as ftse
+import fatf.utils.array.tools as fuat
 import fatf.utils.data.augmentation as fuda
+import fatf.utils.data.datasets as fatf_datasets
 import fatf.utils.data.discretisation as fudd
+import fatf.utils.models as fum
+import fatf.utils.testing.transparency as futt
+import fatf.utils.testing.imports as futi
 
-from fatf.exceptions import IncompatibleModelError, IncorrectShapeError
-
-Index = Union[int, str]
-
-
-ONE_D_ARRAY = np.array([0, 4, 3, 0])
-NUMERICAL_NP_ARRAY_TARGET = np.array([2, 0, 1, 1, 0, 2])
-NUMERICAL_NP_ARRAY = np.array([[0, 0, 0.08, 0.69], [1, 0, 0.03, 0.29],
-                               [0, 1, 0.99, 0.82], [2, 1, 0.73, 0.48],
-                               [1, 0, 0.36, 0.89], [0, 1, 0.07, 0.21]])
-NUMERICAL_STRUCT_ARRAY = np.array([(0, 0, 0.08, 0.69), (1, 0, 0.03, 0.29),
-                                   (0, 1, 0.99, 0.82), (2, 1, 0.73, 0.48),
-                                   (1, 0, 0.36, 0.89), (0, 1, 0.07, 0.21)],
-                                  dtype=[('a', 'i'), ('b', 'i'), ('c', 'f'),
-                                         ('d', 'f')])
-CATEGORICAL_NP_ARRAY = np.array([['a', 'b', 'c'], ['a', 'f', 'g'],
-                                 ['b', 'c', 'c'], ['b', 'f', 'c'],
-                                 ['a', 'f', 'c'], ['a', 'b', 'g']])
-CATEGORICAL_STRUCT_ARRAY = np.array([('a', 'b', 'c'), ('a', 'f', 'g'),
-                                     ('b', 'c', 'c'), ('b', 'f', 'c'),
-                                     ('a', 'f', 'c'), ('a', 'b', 'g')],
-                                    dtype=[('a', 'U1'), ('b', 'U1'), ('c',
-                                                                      'U1')])
-MIXED_ARRAY = np.array([(0, 'a', 0.08, 'a'), (0, 'f', 0.03, 'bb'),
-                        (1, 'c', 0.99, 'aa'), (1, 'a', 0.73, 'a'),
-                        (0, 'c', 0.36, 'b'), (1, 'f', 0.07, 'bb')],
-                       dtype=[('a', 'i'), ('b', 'U1'), ('c', 'f'), ('d',
-                                                                    'U2')])
-
-# Test LIME and bLIMEy on IRIS dataset
 IRIS_DATASET = fatf_datasets.load_iris()
 
 
-class InvalidModel(object):
+def test_sklearn_imports(caplog):
     """
-    Tests for exceptions when a model lacks the ``predict_proba`` method.
[…]
