+++ fatf/__init__.py
@@ -26,7 +26,10 @@ _logger_formatter = logging.Formatter(  # pylint: disable=invalid-name
     datefmt='%y-%b-%d %H:%M:%S')
 _logger_handler.setFormatter(_logger_formatter)
 logger.addHandler(_logger_handler)
-logger.setLevel(logging.INFO)
+if os.environ.get('PYTEST_IN_PROGRESS', None) is None:
+    logger.setLevel(logging.INFO)  # pragma: nocover
+else:
+    logger.setLevel(logging.NOTSET)
 
 # Redirect warnings to the logger module
 # logging.captureWarnings(True)
