+++ pytest.ini
@@ -1,4 +1,5 @@
 [pytest]
+log_level=NOTSET
 filterwarnings =
   error
   default::ImportWarning

[…]
