### 2019-11-02 4f36f4a
CI improvements (#30)
* CI PyPi deployment and RNG seeding
* Python 3.8 testing and doctest fix
* scipy and numpy wheels not available for Python 3.8
* Version added note and Thales URL fix -- The SSL certificate of Thales' website (https://www.thalesgroup.com/en) could not be verified, hence the website URL was removed.
