+++ fatf/__init__.py
@@ -10,11 +10,21 @@ intelligence systems.
 # Author: Kacper Sokol <k.sokol@bristol.ac.uk>
 # License: new BSD
 
+from typing import Optional
+
 import logging
 import os
 import re
 import sys
 import warnings
+
+# Author and license information
+__author__ = 'Kacper Sokol'
+__email__ = 'k.sokol@bristol.ac.uk'
+__license__ = 'new BSD'
+
+# The current package version
+__version__ = '0.0.2-dev'
 
 __all__ = ['setup_warning_filters', 'setup_random_seed']
 
@@ -67,29 +77,82 @@ def setup_warning_filters():
 if 'PYTEST_IN_PROGRESS' not in os.environ:
     setup_warning_filters()  # pragma: no cover
 
-# Set the current package version
-__version__ = '0.0.1'
-
 
 # This function is tested in fatf.tests.test_rngs_seeding
-def setup_random_seed():
+def setup_random_seed(seed: Optional[int] = None) -> None:
     """
     Sets up Python's and numpy's random seed.
 
     Fixture for the tests to assure globally controllable seeding of random
     number generators in both Python (:func:`random.seed`) and ``numpy``
     (``numpy.random.seed``). The seed is taken either from ``FATF_SEED``
-    system variable; if not given it's sampled uniformly from range
-    0--2147483647.
+    system variable or from the ``seed`` input parameter; if neither of
+    the two is given, it is sampled uniformly from 0--2147483647 range.
+
+    .. note::
+
+       If both ``FATF_SEED`` system variable and ``seed`` input parameter are
+       given, the ``seed`` parameter takes the precedence.
+
+    This function loggs (``info``) the origin of the random seed and its value.
+
+    Parameters
+    ----------
+    seed : integer, optional (default=None)
+        .. versionadded:: 0.0.2
+
+        An integer in 0--2147483647 range used to seed Python's and numpy's
+        random number generator.
+
+    Raises
+    ------
+    TypeError
+        The ``seed`` input parameter is not an integer.
+    ValueError
+        The ``seed`` input parameter is outside of the allowed 0--2147483647
+        range. The random seed retrieved from the ``FATF_SEED`` system variable
+        is either outside of the allowed range or cannot be parsed as an
+        integer.
     """
     import numpy as np
     import random
 
-    # It could have been provided in the environment
-    _random_seed = os.environ.get('FATF_SEED', None)
-    if _random_seed is None:
-        _random_seed = np.random.uniform() * (2**31 - 1)
-    _random_seed = int(_random_seed)
+    lower_bound = 0
+    upper_bound = 2147483647
+
+    if seed is None:
+        # It could have been provided in the environment
+        _random_seed_os = os.environ.get('FATF_SEED', None)
+        if _random_seed_os is not None:
+            # Random seed given as a system variable
+            _random_seed_os = _random_seed_os.strip()
+            if _random_seed_os.isdigit():
+                _random_seed = int(_random_seed_os)
+                if _random_seed < lower_bound or _random_seed > upper_bound:
+                    raise ValueError('The random seed retrieved from the '
+                                     'FATF_SEED system variable ({}) is '
+                                     'outside of the allowed 0--2147483647 '
+                                     'range.'.format(_random_seed))
+                logger.info('Seeding RNGs using the system variable.')
+            else:
+                raise ValueError('The random seed retrieved from the '
+                                 'FATF_SEED system variable ({}) '
+                                 'cannot be parsed as a non-negative '
+                                 'integer.'.format(_random_seed_os))
+        else:
+            # No user-defined random seed -- generate randomly
+            _random_seed = int(np.random.uniform() * (2**31 - 1))
[…]
