+++ .travis.yml
@@ -117,8 +117,9 @@ jobs:
       if: |
         env(FATF_TEST_SCIPY) = 1.3.0 AND \
         env(FATF_TEST_NUMPY) = 1.16.3
-      install: skip
+      install: make install-sans-dep
       script: skip
+      before_deploy: bash build_tools/pypi/version-validation.sh
       deploy:
         provider: pypi
         user: "__token__"
@@ -126,7 +127,7 @@ jobs:
           secure: "OC0Z4KEyI27x9p2PeohulMh9SWWBGwX2EgK5+rIkjM8Crr/IVZC2UmJnwi4CC2xQHYbB5qRLEQo4lzHXPKiNgEoHubUZo1ZKnlqJI+feyDSvlPRoBrm5aTdWj7sI30kTXYbAYFN1roE+PA1ZVhJ9206stq+BF5MvGoVloy+DuGP3xEMI3iB/AQd4Fu5OKrM0z5mTJMSd2m+mo47gQ+NoyGXNWJmAkMOKA5p9zqpcS9SgSuBUgIY9vAg+huC4Kn56cJRKe8aZlpTyLEYnggvY63MAlaNF2lY006/clwRUYgNh56fejecJFMdcdCqwPTSJiPfJDQsnwqqQ3QW8NlPFne2Ct7nKFWrpmC9NkbkLRSapcADmeTBpucXZt2F78wS1yagCLcCVYRRhs0bWswmbNVSERuaNo/nZ3kiOrSwFV3/foHqBwVYUKgmyCxPObf+iYWslPLpam/RuHdeoWaOhvKdl/xgvJw/qUCJoTBC2W/Z+OVjNPYmk/PtzcQ8CZ3jRfYg0pyr9zFSjMfGCVRpVU7or4muOC9MtVH8eeGsZyVj3CgWc85/JF/G6MHF0OuJDbBsWHsSekLKSjCJy3aCpJSyuOw+VHyUDyTslCLAc5Lh5PQkVK23lYjrDWVjSPD3Y6pFxQLq+SI3hLfvMoZvtfx9QHFPq++Gh9FQMixFtYX0="
         distributions: "sdist bdist_wheel"
         skip_existing: true
-        skip_cleanup: false
+        skip_cleanup: true
         on:
           <<: *deploy-conditions
           tags: true
@@ -136,9 +137,13 @@ jobs:
       if: |
         env(FATF_TEST_SCIPY) = 1.3.0 AND \
         env(FATF_TEST_NUMPY) = 1.16.3
-      install: pip install --upgrade setuptools wheel
+      install:
+        - pip install --upgrade setuptools wheel
+        - make install-sans-dep
       script: skip
-      before_deploy: make build
+      before_deploy:
+        - bash build_tools/pypi/version-validation.sh
+        - make build
       deploy:
         provider: releases
         api_key:
