### 2019-11-02 4ae84ff
RNGs seeding logging fix
