### 2019-11-02 bc72e74
Linear models testing fix
