+++ fatf/utils/data/instance_augmentation.py
@@ -6,7 +6,8 @@ augmentation function for 1-dimensional numpy array-like objects.
 #         Kacper Sokol <k.sokol@bristol.ac.uk>
 # License: new BSD
 
-from typing import Tuple, Union
+from typing import Tuple  # pylint: disable=unused-import
+from typing import Union
 
 import numpy as np
 
