+++ fatf/utils/data/augmentation.py
@@ -9,7 +9,8 @@ The :mod:`fatf.utils.data.augmentation` module implements data set augmenters.
 # pylint: disable=too-many-lines
 
 from numbers import Number
-from typing import Callable, List, Optional, Set, Tuple, Union
+from typing import Callable, List, Optional, Tuple, Union
+from typing import Set  # pylint: disable=unused-import
 
 import abc
 import logging
