+++ fatf/transparency/sklearn/tests/test_linear_model_sklearn.py
@@ -18,20 +18,27 @@ import pytest
 
 import numpy as np
 
+import fatf
+
 import fatf.transparency.sklearn.linear_model as ftsl
 
 # yapf: disable
 LINEAR_CLASSIFIERS = [
     sklearn.svm.LinearSVC,
-    sklearn.linear_model.SGDClassifier,
-    sklearn.linear_model.PassiveAggressiveClassifier,
     sklearn.linear_model.RidgeClassifier,
     sklearn.linear_model.RidgeClassifierCV,
-    sklearn.linear_model.Perceptron,
     sklearn.discriminant_analysis.LinearDiscriminantAnalysis,
     #
     sklearn.linear_model.LogisticRegression,
-    sklearn.linear_model.LogisticRegressionCV]
+    sklearn.linear_model.LogisticRegressionCV
+]
+# These three linear models get different results in different Python versions,
+# hence will not be tested for parameters.
+LINEAR_CLASSIFIERS_ = [
+    sklearn.linear_model.SGDClassifier,
+    sklearn.linear_model.PassiveAggressiveClassifier,
+    sklearn.linear_model.Perceptron
+]
 LINEAR_REGRESSORS = [
     sklearn.linear_model.BayesianRidge,
     sklearn.linear_model.ElasticNet,
@@ -55,12 +62,14 @@ LINEAR_REGRESSORS = [
     #
     sklearn.svm.LinearSVR,
     sklearn.linear_model.PassiveAggressiveRegressor,
-    sklearn.linear_model.SGDRegressor]
+    sklearn.linear_model.SGDRegressor
+]
 LINEAR_MULTITASK_REGRESSORS = [
     sklearn.linear_model.MultiTaskLasso,
     sklearn.linear_model.MultiTaskLassoCV,
     sklearn.linear_model.MultiTaskElasticNet,
-    sklearn.linear_model.MultiTaskElasticNetCV]
+    sklearn.linear_model.MultiTaskElasticNetCV
+]
 NON_LINEAR_MODELS = [
     sklearn.linear_model.RANSACRegressor,
     sklearn.cluster.KMeans,
@@ -70,7 +79,8 @@ NON_LINEAR_MODELS = [
     sklearn.neighbors.KNeighborsClassifier,
     sklearn.neighbors.KNeighborsRegressor,
     sklearn.tree.DecisionTreeClassifier,
-    sklearn.tree.DecisionTreeRegressor]
+    sklearn.tree.DecisionTreeRegressor
+]
 
 LINEAR_REG_COEF = [
     np.array([0.001, -0.003, 0.004, -0.008]),
@@ -96,13 +106,20 @@ LINEAR_REG_COEF = [
     np.array([0.017, -0.003, 0.040, -0.005]),
     np.array([1.219, 10.356, -0.982, -19.025])  # / 1e+10
 ]
-LINEAR_CLF_COEF = [
-    np.array([[0.004, -0.003, 0.013, -0.033]]),
+LINEAR_CLF_COEF_36 = [
     np.array([[-28.064, -84.191, -65.482, -299.345]]),
     np.array([[0.001, -0.007, 0.004, -0.042]]),
+    np.array([[-3., -9., -7., -32.]]),
+]
+LINEAR_CLF_COEF_37 = [
+    np.array([[84.191, 9.355, 130.964, -514.500]]),
+    np.array([[0.003, -0.003, 0.008, -0.038]]),
+    np.array([[9., 1., 14., -55.]])
+]
+LINEAR_CLF_COEF = [
+    np.array([[0.004, -0.003, 0.013, -0.033]]),
     np.array([[0.065, -0.007, 0.039, -0.011]]),
[…]
