### 2019-11-02 fc9b0d4
Linear model test fix
