+++ fatf/transparency/sklearn/tests/test_linear_model_sklearn.py
@@ -102,7 +102,7 @@ LINEAR_REG_COEF = [
     np.array([0.033, -0.003, 0.019, -0.005]),
     np.array([0.021, -0.003, 0.017, -0.006]),
     #
-    np.array([0.011, 0.007, 0.022, -0.001]),
+    np.array([0.012, 0.007, 0.027, -0.016]),
     np.array([0.017, -0.003, 0.040, -0.005]),
     np.array([1.219, 10.356, -0.982, -19.025])  # / 1e+10
 ]
