### 2019-11-02 107417e
Linear model test fix
