### 2019-11-03 fb4dc56
Version added/changed documentation + API
