+++ fatf/transparency/predictions/surrogate_explainers.py
@@ -1,6 +1,8 @@
 """
-The :mod:`fatf.transparency.surrogate_explainers` module implements example
-surrogate explainers.
+.. versionadded:: 0.0.2
+
+The :mod:`fatf.transparency.predictions.surrogate_explainers` module implements
+example surrogate explainers.
 
 Guidelines and tips for building custom surrogate explainers for various
 types of data (tabular, image and text) can be found in the how_to_guide_
@@ -44,7 +46,9 @@ import fatf.utils.kernels as fatf_kernels
 import fatf.utils.models.validation as fumv
 import fatf.utils.models.models as fumm
 
-__all__ = ['TabularBlimeyLime', 'TabularBlimeyTree']
+__all__ = ['SurrogateTabularExplainer',
+           'TabularBlimeyLime',
+           'TabularBlimeyTree']  # yapf: disable
 
 logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
 
@@ -86,9 +90,8 @@ def _input_is_valid(
     Validates the input parameters of the ``SurrogateTabularExplainer`` class.
 
     For the input parameters description, warnings and exceptions please see
-    the documentation of the
-    :class:`fatf.transparency.surrogate_explainers.SurrogateTabularExplainer`
-    class.
+    the documentation of the :class:`fatf.transparency.predictions.\
+surrogate_explainers.SurrogateTabularExplainer` class.
 
     Returns
     -------
@@ -224,12 +227,15 @@ class SurrogateTabularExplainer(abc.ABC):
     """
     An abstract parent class for implementing surrogate explainers.
 
+    .. versionadded:: 0.0.2
+
     An abstract class that all surrogate explainer classes should inherit from.
     It contains an ``__init__`` method and an input validator --
     ``_explain_instance_input_is_valid`` -- for the abstract
     ``explain_instance`` method. The validation of the input parameters passed
     to the ``__init__`` method is done via the
-    ``fatf.transparency.surrogate_explainers._input_is_valid`` function.
+    ``fatf.transparency.predictions.surrogate_explainers._input_is_valid``
+    function.
 
     If the ``predictive_model`` is a non-probabilistic classifier
     (``as_probabilistic=False``), it is advised to specify both
@@ -631,7 +637,7 @@ class SurrogateTabularExplainer(abc.ABC):
         This function checks the validity of the ``data_row``.
 
         For the description of exceptions raised by this method please see
-        the documentation of the :func:`fatf.transparency.\
+        the documentation of the :func:`fatf.transparency.predictions.\
 surrogate_explainers.SurrogateTabularExplainer.explain_instance` method.
 
         Parameters
@@ -721,6 +727,8 @@ class TabularBlimeyLime(SurrogateTabularExplainer):
     """
     A tabular LIME explainer -- a surrogate explainer based on a linear model.
 
+    .. versionadded:: 0.0.2
+
     This class implements Local Interpretable Model-agnostic Explanations
     (LIME_) introduced by [RIBEIRO2016WHY]_. This implementation mirrors the
     one in the `official LIME package`_, which is available under the
@@ -799,8 +807,8 @@ class TabularBlimeyLime(SurrogateTabularExplainer):
     how_to_tabular_surrogates_ *how-to guide*.
 
     For additional parameters, warnings and errors description please see the
-    documentation of the parent class
-    :class:`fatf.transparency.surrogate_explainers.SurrogateTabularExplainer`.
+    documentation of the parent class :class:`fatf.transparency.predictions.\
+surrogate_explainers.SurrogateTabularExplainer`.
 
     .. _LIME: https://github.com/marcotcr/lime
     .. _`official LIME package`: https://github.com/marcotcr/lime
[…]
