+++ fatf/utils/transparency/__init__.py
@@ -1,4 +1,6 @@
 """
+.. versionadded:: 0.0.2
+
 The :mod:`fatf.utils.transparency` module holds transparency utilities.
 """
 # Author: Kacper Sokol <k.sokol@bristol.ac.uk>
