+++ fatf/__init__.py
@@ -24,7 +24,7 @@ __email__ = 'k.sokol@bristol.ac.uk'
 __license__ = 'new BSD'
 
 # The current package version
-__version__ = '0.0.2-dev'
+__version__ = '0.0.2'
 
 __all__ = ['setup_warning_filters', 'setup_random_seed']
 
