### 2019-11-04 f76a9a6
Tabular Surrogates (#29 and #26)
Note: this is a duplicate of #26 and it closes #26 as well.

* Reformatted object validation for use in explainer validation.
* Added submodular pick plus testing. No testing for SKLearnExplainer as not sure what we're doing with it.
* Added base dicretizer and similar tests used for testing augmentation.
* Finished testing and documentation for abstract discretizer class
* Added quartile discretizer and some basic testing.
* Added additional tests and finished quartile discretizer.
* Added making an array binary based off if the values are equal to row.
* Adde
[…]
