+++ fatf/utils/data/tests/test_instance_augmentation.py
@@ -1,0 +1,139 @@
+"""
+Tests instance augmentation functions.
+"""
+# Author: Alex Hepburn <ah13558@bristol.ac.uk>
+#         Kacper Sokol <k.sokol@bristol.ac.uk>
+# License: new BSD
+
+import numpy as np
+
+import pytest
+
+import fatf
+import fatf.utils.array.validation as fuav
+import fatf.utils.data.instance_augmentation as fudi
+
+from fatf.exceptions import IncorrectShapeError
+
+# yapf: disable
+NUMERICAL_NP_ARRAY = np.array([[0, 0, 0.08, 0.69], [0, 1, 0.07, 0.21]])
+NUMERICAL_STRUCT_ARRAY = np.array(
+    [(0, 0, 0.08, 0.69), (1, 0, 0.03, 0.29)],
+    dtype=[('a', 'i'), ('b', 'i'), ('c', 'f'), ('d', 'f')]
+)
+CATEGORICAL_NP_ARRAY = np.array([['a', 'b', 'c'], ['a', 'f', 'g']])
+CATEGORICAL_STRUCT_ARRAY = np.array(
+    [('a', 'b', 'c'), ('a', 'f', 'g')],
+    dtype=[('a', 'U1'), ('b', 'U1'), ('c', 'U1')]
+)
+MIXED_ARRAY = np.array(
+    [(0, 'a', 0.08, 'a'), (0, 'f', 0.03, 'bb')],
+    dtype=[('a', 'i'), ('b', 'U1'), ('c', 'f'), ('d', 'U2')])
+# yapf: enable
+
+
+def test_validate_input():
+    """
+    Tests :func:`fatf.utils.data.instance_augmentation._validate_input`.
+    """
+    shape_msg = ('The data_row must either be a 1-dimensional numpy array or '
+                 'a numpy void object for structured rows.')
+    samples_msg = 'The samples_number parameter must be a positive integer.'
+    samples_type_msg = 'The samples_number parameter must be an integer.'
+
+    with pytest.raises(IncorrectShapeError) as exin:
+        fudi._validate_input(NUMERICAL_NP_ARRAY, None)
+    assert str(exin.value) == shape_msg
+    with pytest.raises(IncorrectShapeError) as exin:
+        fudi._validate_input(NUMERICAL_STRUCT_ARRAY, None)
+    assert str(exin.value) == shape_msg
+
+    with pytest.raises(ValueError) as exin:
+        fudi._validate_input(NUMERICAL_NP_ARRAY[0], -1)
+    assert str(exin.value) == samples_msg
+    with pytest.raises(ValueError) as exin:
+        fudi._validate_input(NUMERICAL_STRUCT_ARRAY[0], 0)
+    assert str(exin.value) == samples_msg
+
+    with pytest.raises(TypeError) as exin:
+        fudi._validate_input(NUMERICAL_NP_ARRAY[0], 'a')
+    assert str(exin.value) == samples_type_msg
+    with pytest.raises(TypeError) as exin:
+        fudi._validate_input(NUMERICAL_STRUCT_ARRAY[0], 40.)
+    assert str(exin.value) == samples_type_msg
+
+    # Everything OK
+    assert fudi._validate_input(NUMERICAL_NP_ARRAY[0], 50)
+    assert fudi._validate_input(NUMERICAL_STRUCT_ARRAY[0], 50)
+    assert fudi._validate_input(CATEGORICAL_NP_ARRAY[0], 50)
+    assert fudi._validate_input(CATEGORICAL_STRUCT_ARRAY[0], 50)
+    assert fudi._validate_input(MIXED_ARRAY[0], 50)
+
+
+def test_binary_sampler():
+    """
[…]
