### 2019-11-04 0fd975e
0.0.2 documentation update and documentation deployment fix (remove old files)
