+++ build_tools/doc/deploy-doc.sh
@@ -13,12 +13,12 @@ if [[ $PWD != */fat-forensics ]]; then
   exit 1
 fi
 
-KEY_ENC="build_tools/doc/fat-forensics-deploy-key.enc"
-KEY="build_tools/doc/fat-forensics-deploy-key"
-
 DOC_REPO="fat-forensics/fat-forensics-doc"
-DOC_SOURCE="doc/_build/html/*"
+DOC_SOURCE="doc/_build/html/"
 DOC_TARGET="$HOME/fat-forensics-doc"
+
+DOC_DEPLOY_USER="fat-forensics-automator"
+DOC_DEPLOY_URL="https://$DOC_DEPLOY_USER:$FAT_FORENSICS_AUTOMATOR_GITHUB_TOKEN"
 
 DEPLOY_MD5=`git rev-parse --short HEAD`
 if [ -z "$1" ]; then
@@ -30,13 +30,14 @@ fi
 git config user.name "Travis Deploy Bot"
 git config user.email "travis-deploy@fat-forensics.org"
 
-git clone https://fat-forensics-automator:$FAT_FORENSICS_AUTOMATOR_GITHUB_TOKEN@github.com/$DOC_REPO.git --depth 1 $DOC_TARGET
+git clone $DOC_DEPLOY_URL@github.com/$DOC_REPO.git --depth 1 $DOC_TARGET
 
 echo "Pushing the Documentation to GitHub: $DOC_REPO"
-cp -r $DOC_SOURCE $DOC_TARGET
+rsync -rclpDv --exclude=".git" --delete-before $DOC_SOURCE $DOC_TARGET
 cd $DOC_TARGET
 git add -f *
 if [ -f ".nojekyll" ]; then git add -f .nojekyll; fi
+if [ -f ".buildinfo" ]; then git add -f .buildinfo; fi
 
 # Check if anything has actually changed
 GIT_CHANGES=`git status --porcelain`
@@ -44,6 +45,8 @@ GIT_CHANGES=`git status --porcelain`
 if [ -z "$GIT_CHANGES" ]; then
   echo "Nothing to commit. No documentation deployment necessary."
 else
+  git status
   git commit -am "Deploying documentation from $DEPLOY_BRANCH:$DEPLOY_MD5"
   git push origin master
+  git status
 fi
