+++ doc/getting_started/cite.rst
@@ -11,33 +11,43 @@ Citing the Package
 ==================
 
 When referring to the package, any of its parts, architecture or
-functionality in scientific literature, please reference:
+functionality in scientific literature, please reference "*FAT Forensics: A \
+Python Toolbox for Algorithmic Fairness, Accountability and Transparency*",
+which is available on arXiv_.
 
 .. code:: latex
 
    @article{sokol2019fat-forensics,
-     title={FAT Forensics: A Python Toolbox for Algorithmic Fairness,
-            Accountability and Transparency},
+     title={{FAT} {F}orensics: {A} {P}ython {T}oolbox for {A}lgorithmic
+            {F}airness, {A}ccountability and {T}ransparency},
      author={Sokol, Kacper and Santos-Rodriguez, Raul and Flach, Peter},
      journal={arXiv preprint arXiv:1909.05167},
+     url={https://arxiv.org/abs/1909.05167},
      year={2019}
    }
+
+.. _arXiv: https://arxiv.org/abs/1909.05167
 
 Citing the Code
 ===============
 
-If you want to reference any part of the API, you may consider using the
-following reference instead:
+If you want to reference any part of the API (mainly from the software
+engineering perspective) or the code repository itself, you may consider
+referencing "*FAT Forensics: A Python Toolbox for Implementing and Deploying \
+Fairness, Accountability and Transparency Algorithms in Predictive Systems*"
+instead.
 
 .. code:: latex
 
    @article{sokol2019fatf,
-     title={FAT Forensics: A Python Toolbox for Implementing and Deploying
-            Fairness, Accountability and Transparency Algorithms in Predictive
-            Systems},
+     title={{FAT} {F}orensics: {A} {P}ython {T}oolbox for {I}mplementing and
+            {D}eploying {F}airness, {A}ccountability and {T}ransparency
+            {A}lgorithms in {P}redictive {S}ystems},
      author={Sokol, Kacper and Hepburn, Alexander and Poyiadzi, Rafael and
              Clifford, Matthew and Santos-Rodriguez, Raul and Flach, Peter},
-     journal={arXiv preprint},
+     journal={Journal of Open Source Software (JOSS) -- under review},
+     url={https://github.com/openjournals/joss-papers/blob/joss.01727/joss.01727/10.21105.joss.01727.pdf},
+     note={https://github.com/openjournals/joss-reviews/issues/1727},
      year={2019}
    }
 
