+++ README.rst
@@ -97,19 +97,23 @@ FAT Forensics requires **Python 3.5** or higher and the following dependencies:
 
 In addition, some of the modules require *optional* dependencies:
 
-+----------------------------------------+------------------+------------+
-| ``fatf`` module                        | Package          | Version    |
-+========================================+==================+============+
-| ``fatf.transparency.lime``             |                  |            |
-+----------------------------------------+                  |            |
-| ``fatf.transparency.models.lime``      | LIME_            | >=0.0.0.0  |
-+----------------------------------------+                  |            |
-| ``fatf.transparency.predictions.lime`` |                  |            |
-+----------------------------------------+------------------+------------+
-| ``fatf.transparency.sklearn``          | `scikit-learn`_  | >=0.19.2   |
-+----------------------------------------+------------------+------------+
-| ``fatf.vis``                           | matplotlib_      | >=3.0.0    |
-+----------------------------------------+------------------+------------+
++--------------------------------------------------------+------------------+------------+
+| ``fatf`` module                                        | Package          | Version    |
++========================================================+==================+============+
+| ``fatf.transparency.lime``                             |                  |            |
++--------------------------------------------------------+                  |            |
+| ``fatf.transparency.models.lime``                      | LIME_            | >=0.0.0.0  |
++--------------------------------------------------------+                  |            |
+| ``fatf.transparency.predictions.lime``                 |                  |            |
++--------------------------------------------------------+------------------+------------+
+| ``fatf.transparency.predictions.surrogate_explainers`` |                  |            |
++--------------------------------------------------------+                  |            |
+| ``fatf.transparency.sklearn``                          | `scikit-learn`_  | >=0.19.2   |
++--------------------------------------------------------+                  |            |
+| ``fatf.utils.data.feature_selection.sklearn``          |                  |            |
++--------------------------------------------------------+------------------+------------+
+| ``fatf.vis``                                           | matplotlib_      | >=3.0.0    |
++--------------------------------------------------------+------------------+------------+
 
 User Installation
 -----------------

[…]
