### 2019-11-08 #1727 @arokem
:wave: @linuxscout : would you be willing to review this paper for JOSS?
