### 2019-11-20 #1727 @arokem
👋 osolari : would you be willing to review this paper for JOSS?
