### 2019-11-20 #1727 @arokem
:wave: @professoralkmin: would you be willing to review this paper for JOSS?
