### 2019-11-20 #1727 @arokem
Thanks! The reviewer guidelines are [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
