### 2019-11-20 #1727 @professoralkmin
would be happy to do it! :)

Em qua, 20 de nov de 2019 3:56 PM, Ariel Rokem <notifications@github.com>
escreveu:

> 👋@osolari <https://github.com/osolari> : would you be willing to review
> this paper for JOSS?
>
