### 2019-12-05 #1904 @danielskatz
👋 @arfon - something seems messed up here - the reference that pandoc mentions isn't in the paper, I don't think, but is one of the reviewers (?!)  And the reviewer info in the first comment in this issue seems to have an extra ```@``` - Any ideas?
