### 2019-12-05 #1904 @arfon
> 👋 @arfon - something seems messed up here - the reference that pandoc mentions isn't in the paper, I don't think, but is one of the reviewers (?!) And the reviewer info in the first comment in this issue seems to have an extra @ - Any ideas?

Very weird. I'm not sure what's going on here. I've manually fixed up the review issue at the top. Also, the compilation error is related to the extra `@` which was then being fed into Pandoc (as we now include reviewer names in the compiled papers).
