### 2019-12-26 #1904 @osolari
Hi, I wasn't able to install the package. Did others manage to install it using simple pip?
