### 2019-12-26 #1904 @osolari
@bernease I finished my review, but could not install the package using pip. Have you given it a try?
