### 2019-12-26 #1904 @So-Cool
@osolari, have you tried `pip install fat-forensics`?
