### 2019-12-26 #1904 @So-Cool
@osolari, could you please paste here the log of executing this command?
