### 2019-12-26 #1904 @osolari
ERROR: Could not find a version that satisfies the requirement fat-forensics (from versions: none)
ERROR: No matching distribution found for fat-forensics
