### 2019-12-26 #1904 @osolari
nvm. I tried it on my own machine and it was installed. sorry!
