### 2019-12-26 #1904 @So-Cool
Are you trying to install it on Windows? At the moment only Linux and MacOS are supported. You may try to install it directly from sources (which may or may not work):
```
pip install git+https://github.com/fat-forensics/fat-forensics.git
```
