### 2019-12-27 #1904 @arokem
Hi @osolari,  Thanks for your review! I see that there are still two boxes left unchecked in your review (functionality and references). Is this because you still have issues with these two items? If these are not properly addressed right now, could you please make some comments about the changes you believe need to be made?
