### 2019-12-27 #1904 @osolari
Sorry I forgot to check functionality.

About the other two though. I think a brief literature review of the field with more references would help me understand the importance of the work better. Currently the report only references FATFORENSIC itself.
