### 2020-01-02 #1904 @arokem
Hey @osolari : have your previously-raised concerns ([here](https://github.com/openjournals/joss-reviews/issues/1904#issuecomment-569320507)) been addressed?
