+++ fatf/accountability/models/background_check.py
@@ -1,0 +1,22 @@
+
+def background_posterior(scores, mu0, mu1, dens_min, dens_max):
+    return bg_fg_poseriors
+
+
+def update_posterior(bg_fg_posteriors, class_posteriors):
+    return updated_posteriors
+
+class BackgroundCheck(object):
+    def __init__(self, clf):
+        # Check that clf is probabilistic
+        density_estimator = initialize_svm()
+        pass
+
+
+    def fit(self, X):
+        self.density_estimator.fit(X)
+        pass
+
+    def predict_proba(self, X, mu0=1.0, mu1=0.0)
+        pass
+
