### 2020-01-21 #1904 @bernease
Sorry for letting this fall off my radar. I can have the feedback into you by EOD Thursday conservatively, perhaps earlier if I'm able to clear some time later today.
