### 2020-01-22 0f669fd
Removed abstract class - Changed the inheritance fro Abstract class to object while it is not implemented
