### 2020-01-24 #1904 @bernease
I have completed the review to the best of my ability.

I have some concerns about the lack of references to related work as well as discussion of related software in the field, leading me keep these two fields unchecked for now. There is a singular reference to a pre-print manuscript describing the FATForensics software package in depth, Sokol, Santos-Rodriguez, and Flach (2019).

At minimum, would like to see directly comparative software packages in Python (e.g., eli5, AI Fairness 360) referenced and compared to FATForensics in JOSS submission. As the article stands, it is difficult to see where this work fits into the landscape of software without consulting the other manuscript from the authors. Should that be sufficient, I will mark these two criteria as met.
