### 2020-01-24 #1904 @So-Cool
Thank you for the review, @bernease. I'll address your and @osolari's comments over the weekend and update the paper. I'll make sure to let you know when this is done.
