+++ fatf/accountability/models/background_check.py
@@ -1,16 +1,12 @@
 import numpy as np
-from scipy.special import expit
-from sklearn.svm import OneClassSVM
 
 
-def background_posterior(scores, mu0, mu1, delta, max_dens):
-    # TODO maybe apply expit only on the necessary cases
-    p_x_and_f = expit(scores + delta)
-    # TODO look for other methods to clip the probabilities?
-    q = np.clip(p_x_and_f / max_dens, 0.0, 1.0)
+def background_posterior(relative_densities, mu0, mu1, density_range=[0., 1.]):
+	# TODO We could change the clipping for an assert
+    q = np.clip(relative_densities, density_range[0], density_range[1])
     p_x_and_b = q * mu1 + (1.0 - q) * mu0
 
-    bg_fg_posteriors = np.zeros((len(scores), 2))
+    bg_fg_posteriors = np.zeros((len(relative_densities), 2))
     bg_fg_posteriors[:, 0] = p_x_and_b / (p_x_and_b + q)
     bg_fg_posteriors[:, 1] = 1.0 - bg_fg_posteriors[:, 0]
 
@@ -24,29 +20,18 @@ def update_posterior(bg_fg_posteriors, class_posteriors):
 
 
 class BackgroundCheck(object):
-    def __init__(self, clf, density_estimator=OneClassSVM()):
-        # Check that clf is probabilistic
+    def __init__(self, clf, rde):
+        '''
+            clf: pretrained probabilistic classifier
+                Per each samples always gives K probabilities that sum to one,
+                where K is the number of classes.
+                clf requires function predict_proba
+            rde: pretrained relative density estimator
+                Per each sample always gives one score between 0 and 1
+        '''
         self._clf = clf
-        self._density_estimator = density_estimator
-
-
-    def fit(self, X):
-        """Fits the density estimator to the data in X.
-
-        Args:
-            X (array-like, shape = [n_samples, n_features]): training data.
-
-        Returns:
-            Nothing.
-
-        """
-        self._density_estimator.fit(X)
-        surrogate_densities = self._density_score(X)
-        # Minimum density from the training set
-        self._delta = - surrogate_densities.min()
-        # Sets the 0.5 probability to the sample with minimum density
-        relative_densities = expit(surrogate_densities + self._delta)
-        self._max_dens  = relative_densities.max()
+        self._rde = rde
+        return self
 
 
     def predict_proba(self, X, mu0=1.0, mu1=0.0):
@@ -61,32 +46,12 @@ class BackgroundCheck(object):
             probabilities for background (column 0) and foreground (column 1).
 
         """
-        surrogate_densities = self._density_score(X)
-        bg_fg_posteriors = background_posterior(surrogate_densities, mu0, mu1,
-                                                self._delta, self._max_dens)
+        relative_densities = self._rde.score(X)
+
+        bg_fg_posteriors = background_posterior(relative_densities, mu0, mu1)
 
         class_posteriors = self._clf.predict_proba(X)
 
         class_bg_posteriors = update_posterior(bg_fg_posteriors,
                                                class_posteriors)
         return class_bg_posteriors
-
-
-    def _density_score(self, X):
-        """Gets scores for the objects of X using different functions that
-        depend on the estimator.
-
-        Args:
-            X (array-like, shape = [n_samples, n_features]): training data.
-
-        Returns:
-            (array-like, shape = [n_samples]): scores given by the estimator.
-
-        """
-        if 'score_samples' in dir(self._density_estimator):
-            s = self._density_estimator.score_samples(X)
-        elif 'score' in dir(self._density_estimator):
-            s = self._density_estimator.score(X)
-        elif 'decision_function' in dir(self._density_estimator):
-            s = self._density_estimator.decision_function(X).reshape(-1)
-        return s
