+++ fatf/accountability/models/tests/test_background_check.py
@@ -13,6 +13,7 @@ import fatf.utils.models as fatf_models
 from fatf.accountability.models.background_check import BackgroundCheck
 from fatf.accountability.models.background_check import update_posterior
 
+from fatf.utils.data.density import GaussianRelativeDensityEstimator
 
 def test_background_check_pipeline():
     """
@@ -28,9 +29,12 @@ def test_background_check_pipeline():
     clf = fatf_models.KNN()
     clf.fit(X, y)
 
-    # Train Background Check
-    bc = BackgroundCheck(clf)
-    bc.fit(X)
+    # Train a relative density estimator
+    rde = GaussianRelativeDensityEstimator()
+    rde.fit(X)
+
+    # Create Background Check
+    bc = BackgroundCheck(clf, rde)
 
     # Test Background Check
     class_bg_posteriors = bc.predict_proba(X, mu0=1.0, mu1=0.0)
@@ -39,3 +43,5 @@ def test_background_check_pipeline():
     assert(class_bg_posteriors.shape[0] == X.shape[0])
     # One additional class
     assert(class_bg_posteriors.shape[1] == n_classes + 1)
+
+
