### 2020-01-29 3b6a6f8
Resolved requested changes
