+++ doc/JOSS_paper/paper.md
@@ -22,6 +22,7 @@ authors:
   - name: Matthew Clifford
     affiliation: 2
   - name: Raul Santos-Rodriguez
+    orcid: 0000-0001-9576-3905
     affiliation: 2
   - name: Peter Flach
     orcid: 0000-0001-6857-5810
@@ -36,92 +37,143 @@ bibliography: paper.bib
 ---
 
 # Background #
+
 Predictive systems, in particular machine learning algorithms, can take
-important decisions -- sometimes legally binding -- about our everyday life.
-In most cases, however, these systems and decisions are neither regulated nor
-certified. Given the potential harm that these systems can cause, qualities
-such as **fairness**, **accountability** and **transparency** (FAT) of
-predictive systems are of paramount importance. To ensure high-quality, fair,
-transparent and reliable predictive systems, we developed an open source Python
-package called *FAT Forensics* that can inspect important fairness,
-accountability and transparency aspects of these systems to automatically and
-objectively report them back to their engineers and users. Our toolbox offers
-full functionality for evaluating all aspects of a predictive pipeline: data
-(and their features), models and predictions. Published under the BSD 3-Clause
-open source licence, FAT Forensics is opened up for personal and commercial
-usage.
+important, and sometimes legally binding, decisions about our everyday life. In
+most cases, however, these systems and decisions are neither regulated nor
+certified. Given the potential harm that these algorithms can cause, their
+qualities such as **fairness**, **accountability** and **transparency** (FAT)
+are of paramount importance. To ensure high-quality, fair, transparent and
+reliable predictive systems, we developed an open source Python package called
+*FAT Forensics*. It can inspect important fairness, accountability and
+transparency aspects of predictive algorithms to automatically and objectively
+report them back to engineers and users of such systems. Our toolbox can
+evaluate all elements of a predictive pipeline: data (and their features),
+models and predictions. Published under the BSD 3-Clause open source licence,
+*FAT Forensics* is opened up for personal and commercial usage.
 
 # Summary #
-FAT Forensics is designed as an interoperable framework for *implementing*,
-*testing* and *deploying* novel algorithms invented by the FAT research
-community and facilitates their evaluation and comparison against the state of
-the art, hence democratising access to these techniques. In addition to
+
+*FAT Forensics* is designed as an interoperable framework for *implementing*,
+*testing* and *deploying* novel algorithms devised by the FAT research
+community. It facilitates their evaluation and comparison against the state of
+the art, thereby democratising access to these techniques. In addition to
 supporting research in this space, the toolbox is capable of analysing all
-artefacts of a predictive pipeline -- data, models and predictions -- by
+components of a predictive pipeline -- data, models and predictions -- by
 considering their fairness, accountability (robustness, security, safety and
-privacy) and transparency (interpretability and explainability). FAT Forensics
-collates all of these diverse tools and algorithms under a common application
-programming interface (API) using a modular design with shared core algorithmic
-components, thereby making the process of creating new algorithm as easy as
-connecting the right blocks -- see Figure 1.
+privacy) and transparency (interpretability and explainability).
 
-![FAT Forensics modular architecture design.](software_design.png){width=45%}
+*FAT Forensics* collates all of these diverse tools and algorithms under a
+common application programming interface. This is achieved with a modular
+design that allows to share and reuse a collection of core algorithmic
+components -- see Figure 1. This architecture makes the process of creating new
+algorithms as easy as connecting the right blocks, therefore supporting a range
+of diverse use cases.
 
-The input requirements for data sets and predictive models are kept to a
-minimum, lowering any barriers for adoption of FAT Forensics in new and already
[…]
