+++ doc/JOSS_paper/paper.bib
@@ -1,7 +1,93 @@
 @article{sokol2019fatf,
- title={FAT Forensics: A Python Toolbox for Algorithmic Fairness,
-        Accountability and Transparency},
- author={Sokol, Kacper and Santos-Rodriguez, Raul and Flach, Peter},
- journal={arXiv preprint arXiv:1909.05167},
- year={2019}
+  title={FAT Forensics: A Python Toolbox for Algorithmic Fairness,
+         Accountability and Transparency},
+  author={Sokol, Kacper and Santos-Rodriguez, Raul and Flach, Peter},
+  journal={arXiv preprint arXiv:1909.05167},
+  year={2019}
+}
+
+@article{bellamy2018ai,
+  title={AI Fairness 360: An Extensible Toolkit for Detecting,
+         Understanding, and Mitigating Unwanted Algorithmic Bias},
+  author={Rachel K. E. Bellamy and Kuntal Dey and Michael Hind and
+          Samuel C. Hoffman and Stephanie Houde and Kalapriya Kannan and
+          Pranay Lohia and Jacquelyn Martino and Sameep Mehta and
+          Aleksandra Mojsilovic and Seema Nagar and
+          Karthikeyan Natesan Ramamurthy and John Richards and
+          Diptikalyan Saha and Prasanna Sattigeri and Moninder Singh and
+          Kush R. Varshney and Yunfeng Zhang},
+  journal={arXiv preprint arXiv:1810.01943},
+  year={2018}
+}
+
+@inproceedings{ribeiro2016why,
+  title={"Why Should {I} Trust You?": Explaining the Predictions
+         of Any Classifier},
+  author={Marco Tulio Ribeiro and Sameer Singh and Carlos Guestrin},
+  booktitle={Proceedings of the 22nd {ACM} {SIGKDD} International Conference
+             on Knowledge Discovery and Data Mining, San Francisco, CA, USA,
+             August 13-17, 2016},
+  pages={1135--1144},
+  year={2016}
+}
+
+@article{friedman2001greedy,
+  title={Greedy function approximation: a gradient boosting machine},
+  author={Friedman, Jerome H},
+  journal={Annals of statistics},
+  pages={1189--1232},
+  year={2001},
+  publisher={JSTOR}
+}
+
+@inproceedings{agarwal2018reductions,
+  title={A Reductions Approach to Fair Classification},
+  author={Agarwal, Alekh and Beygelzimer, Alina and Dudik, Miroslav and
+          Langford, John and Wallach, Hanna},
+  booktitle={Proceedings of the 35th International Conference
+             on Machine Learning},
+  pages={60--69},
+  year={2018},
+  editor={Dy, Jennifer and Krause, Andreas},
+  volume={80},
+  series={Proceedings of Machine Learning Research},
+  address={Stockholmsm\"assan, Stockholm Sweden},
+  month={10--15 Jul},
+  publisher={PMLR},
+  pdf={http://proceedings.mlr.press/v80/agarwal18a/agarwal18a.pdf},
+  url={http://proceedings.mlr.press/v80/agarwal18a.html}
+}
+
+@article{goldstein2015peeking,
+  title={Peeking inside the black box: Visualizing statistical learning
+         with plots of individual conditional expectation},
+  author={Goldstein, Alex and Kapelner, Adam and Bleich, Justin and
+          Pitkin, Emil},
+  journal={Journal of Computational and Graphical Statistics},
+  volume={24},
+  number={1},
+  pages={44--65},
+  year={2015},
+  publisher={Taylor \& Francis}
+}
+
+@article{nori2019interpretml,
[…]
