### 2020-01-30 29f1e76
Added option to request relative densities
