### 2020-02-06 #1904 @bernease
I have updated (and completed all marks) based on the updates provided by the author.
