### 2020-02-18 #1904 @arokem
@So-Cool : could you please add the DOIs mentioned above?

Two other comments about references:

1. Since scikit learn is mentioned so prominently, it would be appropriate to cite their 2011 JMLR paper about the software: http://www.jmlr.org/papers/v12/pedregosa11a.html

2. What is the difference between this paper and the arXiv preprint referenced? Is that preprint already published elsewhere?

Thanks!
