+++ doc/JOSS_paper/paper.bib
@@ -1,5 +1,5 @@
 @article{sokol2019fatf,
-  title={FAT Forensics: A Python Toolbox for Algorithmic Fairness,
+  title={FAT {F}orensics: A {P}ython Toolbox for Algorithmic Fairness,
          Accountability and Transparency},
   author={Sokol, Kacper and Santos-Rodriguez, Raul and Flach, Peter},
   journal={arXiv preprint arXiv:1909.05167},
@@ -7,7 +7,7 @@
 }
 
 @article{bellamy2018ai,
-  title={AI Fairness 360: An Extensible Toolkit for Detecting,
+  title={AI {F}airness 360: An Extensible Toolkit for Detecting,
          Understanding, and Mitigating Unwanted Algorithmic Bias},
   author={Rachel K. E. Bellamy and Kuntal Dey and Michael Hind and
           Samuel C. Hoffman and Stephanie Houde and Kalapriya Kannan and
@@ -28,16 +28,18 @@
              on Knowledge Discovery and Data Mining, San Francisco, CA, USA,
              August 13-17, 2016},
   pages={1135--1144},
-  year={2016}
+  year={2016},
+  doi={10.18653/v1/n16-3020}
 }
 
 @article{friedman2001greedy,
-  title={Greedy function approximation: a gradient boosting machine},
+  title={Greedy function approximation: A gradient boosting machine},
   author={Friedman, Jerome H},
   journal={Annals of statistics},
   pages={1189--1232},
   year={2001},
-  publisher={JSTOR}
+  publisher={JSTOR},
+  doi={10.1214/aos/1013203451}
 }
 
 @inproceedings{agarwal2018reductions,
@@ -68,7 +70,8 @@
   number={1},
   pages={44--65},
   year={2015},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.1080/10618600.2014.907095}
 }
 
 @article{nori2019interpretml,
@@ -90,4 +93,16 @@
   version={v1.1.2},
   doi={10.5281/zenodo.1423046},
   url={https://doi.org/10.5281/zenodo.1423046}
+}
+
+@article{scikit-learn,
+  title={Scikit-learn: Machine Learning in {P}ython},
+  author={Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V. and
+          Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P. and
+          Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and
+          Cournapeau, D. and Brucher, M. and Perrot, M. and Duchesnay, E.},
+  journal={Journal of Machine Learning Research},
+  volume={12},
+  pages={2825--2830},
+  year={2011}
 }
