+++ doc/index.rst
@@ -53,7 +53,7 @@ or its documentation.
 .. _`FAT-Forensics organisation`: https://github.com/fat-forensics
 .. _`fat-forensics`: https://github.com/fat-forensics/fat-forensics
 .. _`GitHub issue tracker`: https://github.com/fat-forensics/fat-forensics/issues
-.. _gitter: https://gitter.im/fat-forensics/fat-forensics
+.. _gitter: https://gitter.im/fat-forensics
 .. _slack: https://fat-forensics.slack.com
 .. _`mailing list`: https://groups.google.com/forum/#!forum/fat-forensics
 .. _`University of Bristol`: http://www.bristol.ac.uk
