### 2020-02-19 #1904 @So-Cool
Hi @arokem,
I added the missing DOIs and referenced scikit-learn; thank you for the pointers. The arXiv preprint is an extended version of the JOSS paper with usage examples and more thorough discussion of the relevant literature. We will be revising it in the coming months and submitting it to either a journal or a conference.

If any other revisions are needed, just let me know.
