### 2020-05-16 e56d2ce
scikit-learn>=0.23 and numpy>=1.18 compatibility and testing
