+++ .travis.yml
@@ -20,15 +20,18 @@ before_cache:
   - rm -f $HOME/.cache/pip/log/debug.log
 
 python:
+  - '3.8'
   - '3.7'
   - '3.6'
   - '3.5'
-  - '3.8-dev'
+  - '3.9-dev'
 
 env:
   global:
     - MAKEFLAGS='-j 2'  # This allows to parallelise make, e.g. make t1 t2
   matrix:
+    - FATF_TEST_SCIPY=1.4.1  FATF_TEST_NUMPY=1.18.4
+    - FATF_TEST_SCIPY=1.3.3  FATF_TEST_NUMPY=1.17.5
     - FATF_TEST_SCIPY=1.3.0  FATF_TEST_NUMPY=1.16.3
     - FATF_TEST_SCIPY=1.2.0  FATF_TEST_NUMPY=1.16.0
     - FATF_TEST_SCIPY=1.1.0  FATF_TEST_NUMPY=1.15.4
@@ -51,16 +54,24 @@ matrix:
       env: FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.12.1
     - python: '3.7'
       env: FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.10.1 FATF_TEST_MATPLOTLIB=3.0.0
-    - python: '3.8-dev'
+    - python: '3.8'
       env: FATF_TEST_SCIPY=1.0.1  FATF_TEST_NUMPY=1.14.6
-    - python: '3.8-dev'
+    - python: '3.8'
       env: FATF_TEST_SCIPY=0.19.1 FATF_TEST_NUMPY=1.13.3
-    - python: '3.8-dev'
+    - python: '3.8'
       env: FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.12.1
-    - python: '3.8-dev'
+    - python: '3.8'
+      env: FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.10.1 FATF_TEST_MATPLOTLIB=3.0.0
+    - python: '3.9-dev'
+      env: FATF_TEST_SCIPY=1.0.1  FATF_TEST_NUMPY=1.14.6
+    - python: '3.9-dev'
+      env: FATF_TEST_SCIPY=0.19.1 FATF_TEST_NUMPY=1.13.3
+    - python: '3.9-dev'
+      env: FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.12.1
+    - python: '3.9-dev'
       env: FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.10.1 FATF_TEST_MATPLOTLIB=3.0.0
   allow_failures:
-    - python: '3.8-dev'
+    - python: '3.9-dev'
     - env: FATF_TEST_SCIPY=latest FATF_TEST_NUMPY=latest
 
 install:
