+++ fatf/transparency/sklearn/__init__.py
@@ -6,19 +6,15 @@ The :mod:`fatf.transparency.sklearn` module implements scikit-learn explainers.
 # Author: Kacper Sokol <k.sokol@bristol.ac.uk>
 # License: new BSD
 
-import warnings
-
 try:
     import sklearn
 except ImportError:
-    _warning_msg = (  # pylint: disable=invalid-name
+    raise ImportError(
         'scikit-learn (sklearn) Python module is not installed on your '
         'system. You must install it in order to use '
         'fatf.transparency.sklearn functionality. '
         'One possibility is to install scikit-learn alongside this package '
-        'via machine learning dependencies with: pip install fatf[ml].')
-    warnings.warn(_warning_msg, ImportWarning)
+        'via machine learning dependencies with: pip install '
+        'fat-forensics[ml].')
 else:
     del sklearn
-finally:
-    del warnings
