+++ .travis.yml
@@ -55,6 +55,12 @@ matrix:
     - python: '3.7'
       env: FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.10.1 FATF_TEST_MATPLOTLIB=3.0.0
     - python: '3.8'
+      env: FATF_TEST_SCIPY=1.3.0  FATF_TEST_NUMPY=1.16.3
+    - python: '3.8'
+      env: FATF_TEST_SCIPY=1.2.0  FATF_TEST_NUMPY=1.16.0
+    - python: '3.8'
+      env: FATF_TEST_SCIPY=1.1.0  FATF_TEST_NUMPY=1.15.4
+    - python: '3.8'
       env: FATF_TEST_SCIPY=1.0.1  FATF_TEST_NUMPY=1.14.6
     - python: '3.8'
       env: FATF_TEST_SCIPY=0.19.1 FATF_TEST_NUMPY=1.13.3
