+++ fatf/transparency/lime.py
@@ -1,443 +1,0 @@
-"""
-.. deprecated:: 0.0.3
-   This module will be deprecated in FAT Forensics version 0.0.3.
-   Instead of wrapping the lime package a full (modular) version of the
-   LIME surrogate explainer has been implemented -- see the :class:`fatf.\
-transparency.predictions.surrogate_explainers.TabularBlimeyLime` class and the
-   :ref:`how_to_tabular_surrogates` how-to guide for more details.
-
-The :mod:`fatf.transparency.lime` module wraps the LIME_ explainer.
-
-This module implements a generic tabular data LIME explainer that can be
-customised to either explain models (:mod:`fatf.transparency.models.lime`
-module) or predictions (:mod:`fatf.transparency.predictions.lime` module).
-
-**This module requires the lime package to be installed.**
-
-(At the moment, only the tabular data explainer is wrapped. Since LIME is
-slowly being reimplemented withing this package, wrapping anything else
-would be a waste of time.)
-
-.. _LIME: https://github.com/marcotcr/lime
-"""
-# Author: Alex Hepburn <ah13558@bristol.ac.uk>
-#         Kacper Sokol <k.sokol@bristol.ac.uk>
-# License: new BSD
-
-import logging
-import warnings
-
-from typing import Any, Dict, List, Tuple, Union
-
-import numpy as np
-
-import fatf.utils.array.tools as fuat
-import fatf.utils.array.validation as fuav
-import fatf.utils.models.validation as fumv
-
-from fatf.exceptions import IncompatibleModelError, IncorrectShapeError
-
-try:
-    import lime.lime_tabular
-except ImportError:
-    raise ImportError(
-        'Lime package is not installed on your system. You must install it in '
-        'order to use the fatf.transparency.lime module. One possibility is '
-        'to install LIME alongside this package with: pip install '
-        'fat-forensics[lime].')
-
-__all__ = ['Lime']
-
-logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
-
-
-class Lime(object):
-    """
-    Wraps LIME package's ``lime.lime_tabular.LimeTabularExplainer`` class.
-
-    .. warning::
-        Contrarily to the `LIME tabular explainer`__ this wrapper sets the
-        ``sample_around_instance`` parameter to ``True`` meaning that by
-        default it provides a local rather than a global explanation. This can
-        be changed by either providing a ``sample_around_instance`` or a
-        ``local_explanation`` parameter with the first one taking precedence.
-
-    This LIME_ wrapper can be initialised with any of the
-    ``lime.lime_tabular.LimeTabularExplainer`` named parameters. Additionally,
-    one can also pass in any of the
-    ``lime.lime_tabular.LimeTabularExplainer.explain_instance`` method named
-    parameters, which will be saved within the object and used when the local
-    ``explain_instance`` method is called. In case the same parameters are
-    provided both when initialising the object and when explaining an instance
-    the values passed to the latter take precedence.
-
-    In addition to all the named parameters one may decide to specify a model
-    to be used with the explainer -- the model requires ``predict`` method for
-    the regressor mode and ``predict_proba`` method for the classification mode
-    -- or a predictive function (``predict_fn``), which is accessed directly
-    via the explainer. If both are given, the latter takes the precedence.
-
-    For all of the available parameters please consult the
-    `LIME API documentation`__.
-
-    .. warning::
-        Since LIME does not support structured arrays by default the predictive
-        function and the model have to operate on unstructured types. If a
-        structured data or structured data point are passed in they are
-        converted to unstructured types and these are used inside the lime.
-        If your data array is structured, consider using
[…]
