+++ fatf/transparency/tests/test_lime.py
@@ -1,542 +1,0 @@
-"""
-Tests the LIME wrapper.
-"""
-# Author: Alex Hepburn <ah13558@bristol.ac.uk>
-#         Kacper Sokol <k.sokol@bristol.ac.uk>
-# License: new BSD
-
-import pytest
-
-try:
-    import lime
-except ImportError:  # pragma: no cover
-    pytest.skip(
-        'Skipping lime wrapper tests -- lime missing.',
-        allow_module_level=True)
-else:
-    del lime
-
-import importlib
-import sys
-
-import numpy as np
-
-import fatf.transparency.lime as ftl
-import fatf.utils.models as fum
-import fatf.utils.testing.imports as futi
-import fatf.utils.testing.transparency as futt
-
-from fatf.exceptions import IncompatibleModelError, IncorrectShapeError
-
-# yapf: disable
-FUTURE_WARNING = (
-    'The LIME wrapper will be deprecated in FAT Forensics version '
-    '0.0.3. Please consider using the TabularBlimeyLime explainer '
-    'class implemented in the fatf.transparency.predictions.'
-    'surrogate_explainers module instead. Alternatively, you may '
-    'consider building a custom surrogate explainer using the '
-    'functionality implemented in FAT Forensics -- see the *Tabular '
-    'Surrogates* how-to guide for more details.')
-
-SAMPLE = np.array([0, 1, 0.08, 0.54])
-SAMPLE_STRUCT = np.array(
-    [(0, 1, 0.08, 0.54)],
-    dtype=[('a', 'i'), ('b', 'i'), ('c', 'f'), ('d', 'f')])[0]
-CLF = fum.KNN()
-CLF.fit(futt.NUMERICAL_NP_ARRAY, futt.LABELS)
-CLASS_NAMES = ['class0', 'class1', 'class2']
-FEATURE_NAMES = ['feat0', 'feat1', 'feat2', 'feat3']
-
-NUMERICAL_RESULTS = {
-    'class0': [('feat0 <= 0.00', -0.415),
-               ('0.50 < feat1 <= 1.00', -0.280),
-               ('0.07 < feat2 <= 0.22', 0.037),
-               ('0.34 < feat3 <= 0.58', -0.007)],
-    'class1': [('0.50 < feat1 <= 1.00', 0.202),
-               ('0.07 < feat2 <= 0.22', -0.076),
-               ('feat0 <= 0.00', 0.019),
-               ('0.34 < feat3 <= 0.58', -0.018)],
-    'class2': [('feat0 <= 0.00', 0.395),
-               ('0.50 < feat1 <= 1.00', 0.077),
-               ('0.07 < feat2 <= 0.22', 0.039),
-               ('0.34 < feat3 <= 0.58', 0.025)]
-}
-CATEGORICAL_RESULTS = {
-    'class0': [('feat0=0', -0.413),
-               ('feat1=1', -0.282),
-               ('0.07 < feat2 <= 0.22', 0.0366),
-               ('0.34 < feat3 <= 0.58', -0.00717)],
-    'class1': [('feat1=1', 0.2048),
-               ('0.07 < feat2 <= 0.22', -0.0767),
-               ('feat0=0', 0.0179),
-               ('0.34 < feat3 <= 0.58', -0.018)],
-    'class2': [('feat0=0', 0.395),
-               ('feat1=1', 0.077),
-               ('0.07 < feat2 <= 0.22', 0.039),
-               ('0.34 < feat3 <= 0.58', 0.025)]
-}
-REGRESSION_RESULTS = [
-    ('feat0 <= 0.00', 1.332),
-    ('0.50 < feat1 <= 1.00', 0.767),
-    ('0.34 < feat3 <= 0.58', 0.149),
[…]
