+++ fatf/vis/lime.py
@@ -5,40 +5,49 @@ The :mod:`fatf.vis.lime` module visualises tabular LIME explanations.
 #         Kacper Sokol <k.sokol@bristol.ac.uk>
 # License: new BSD
 
+import logging
+
 from numbers import Number
-from typing import Dict, List, Tuple, Union
+from typing import Any, Dict, Set, Union
 
 import matplotlib.pyplot as plt
 
 __all__ = ['plot_lime']
 
-LimeExplanation = List[Tuple[str, float]]
+logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
+
 BlimeyExplanation = Dict[str, float]
 
 
-def plot_lime(
-        lime_explanation: Union[LimeExplanation, Dict[str, LimeExplanation],
-                                Dict[str, BlimeyExplanation]]) -> plt.Figure:
+def plot_lime(surrogate_explanation: Union[BlimeyExplanation,
+                                           Dict[str, BlimeyExplanation]]
+              ) -> plt.Figure:
     """
-    Plots a LIME explanation.
+    Plots an importance-based surrogate explanation, e.g., LIME.
 
-    This plotting function is intended for the
-    :class:`fatf.transparency.lime.Lime`, :class:`fatf.transparency.\
+    This plotting function is intended for the :class:`fatf.transparency.\
 predictions.surrogate_explainers.TabularBlimeyLime` and :class:`fatf.\
-transparency.predictions.surrogate_explainers.TabularBlimeyTree`  explainers.
+transparency.predictions.surrogate_explainers.TabularBlimeyTree` explainers.
+    When multiple explanations are provided, they will share a common y-axis
+    if they use the same set of interpretable features.
 
     Parameters
     ----------
-    lime_explanation : Dictionary[string, List[Tuple[string, float]]] or \
-Dictionary[string, Dictionary[string, float]] or List[Tuple[string, float]]
-        An explanation returned by the ``explain_instance`` method of the LIME
-        explainer or one of the surrogate explainers. For a classifier this
-        will be a dictionary where the keys are class names and the values are
-        either lists of 2-tuples or dictionaries where the first element (key)
-        is an explanatory feature name and the second element (value) is the
-        importance of this explanatory feature. For regressor explanations this
-        will be simply a list of 2-tuples of the same structure as for the
-        classifier.
+    surrogate_explanation : Dictionary[string, float] or \
+Dictionary[string, Dictionary[string, float]]
+        An explanation returned by the ``explain_instance`` method of a
+        surrogate explainer. For a classifier this will be a dictionary where
+        the keys are class names and the values are dictionaries where the key
+        is an interpretable feature name and the value is the importance of
+        this interpretable feature. For regressor explanations this will be
+        a dictionary where the key is an interpretable feature name and the
+        value is the importance of this interpretable feature.
+
+        .. versionchanged:: 0.1.0
+           Dropped support for LIME explanation format:
+           ``List[Tuple[string, float]]`` for regressors and
+           ``Dictionary[string, List[Tuple[string, float]]]`` for probabilistic
+           classifiers.
 
         .. versionchanged:: 0.0.2
            Support for surrogate explainer explanations of the form:
@@ -47,65 +56,61 @@ Dictionary[string, Dictionary[string, float]] or List[Tuple[string, float]]
     Raises
     ------
     TypeError
-        The ``lime_explanation`` parameter is not a list (regression) or a
-        dictionary (classification). One of the class names is not a string.
-        One of the explanatory features is not a string. One of the explanatory
-        values is not a number.
+        The ``surrogate_explanation`` parameter is not a dictionary. One of the
+        class names is not a string. One of the interpretable feature names is
[…]
