+++ fatf/transparency/sklearn/linear_model.py
@@ -19,7 +19,7 @@ import fatf.utils.tools as fut
 
 __all__ = ['linear_classifier_coefficients', 'SKLearnLinearModelExplainer']
 
-_SKLEARN_VERSION = [int(i) for i in sklearn.__version__.split('.')]
+_SKLEARN_VERSION = [int(i) for i in sklearn.__version__.split('.')[:2]]
 _SKLEARN_0_22 = fut.at_least_verion([0, 22], _SKLEARN_VERSION)
 
 if _SKLEARN_0_22:  # pragma: nocover
