+++ .travis.yml
@@ -37,10 +37,10 @@ env:
     - FATF_TEST_SCIPY=1.1.0  FATF_TEST_NUMPY=1.15.4
     - FATF_TEST_SCIPY=1.0.1  FATF_TEST_NUMPY=1.14.6
     - FATF_TEST_SCIPY=0.19.1 FATF_TEST_NUMPY=1.13.3
-    - FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.12.1
+    - FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.12.1 FATF_TEST_SKLEARN=0.19.2
     # The matplotlib version is fixed to 3.0.0 because since version 3.1 it
     # requires at least 1.11 numpy and we support numpy 1.10
-    - FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.10.1 FATF_TEST_MATPLOTLIB=3.0.0
+    - FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.10.1 FATF_TEST_SKLEARN=0.20.4 FATF_TEST_MATPLOTLIB=3.0.0
     - FATF_TEST_SCIPY=latest FATF_TEST_NUMPY=latest
 
 matrix:
@@ -51,9 +51,9 @@ matrix:
     - python: '3.7'
       env: FATF_TEST_SCIPY=0.19.1 FATF_TEST_NUMPY=1.13.3
     - python: '3.7'
-      env: FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.12.1
+      env: FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.12.1 FATF_TEST_SKLEARN=0.19.2
     - python: '3.7'
-      env: FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.10.1 FATF_TEST_MATPLOTLIB=3.0.0
+      env: FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.10.1 FATF_TEST_SKLEARN=0.20.4 FATF_TEST_MATPLOTLIB=3.0.0
     - python: '3.8'
       env: FATF_TEST_SCIPY=1.3.0  FATF_TEST_NUMPY=1.16.3
     - python: '3.8'
@@ -65,17 +65,17 @@ matrix:
     - python: '3.8'
       env: FATF_TEST_SCIPY=0.19.1 FATF_TEST_NUMPY=1.13.3
     - python: '3.8'
-      env: FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.12.1
+      env: FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.12.1 FATF_TEST_SKLEARN=0.19.2
     - python: '3.8'
-      env: FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.10.1 FATF_TEST_MATPLOTLIB=3.0.0
+      env: FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.10.1 FATF_TEST_SKLEARN=0.20.4 FATF_TEST_MATPLOTLIB=3.0.0
     - python: '3.9-dev'
       env: FATF_TEST_SCIPY=1.0.1  FATF_TEST_NUMPY=1.14.6
     - python: '3.9-dev'
       env: FATF_TEST_SCIPY=0.19.1 FATF_TEST_NUMPY=1.13.3
     - python: '3.9-dev'
[…]
