### 2020-05-16 51bcbbd
scikit-learn>=0.22 pylint fix when an older version is installed
