+++ fatf/transparency/predictions/surrogate_explainers.py
@@ -85,14 +85,15 @@ else:
 BinSamplingValues = Dict[Union[str, int],
                          Dict[int, Tuple[float, float, float, float]]]
 Explanation = Dict[str, Dict[str, float]]
+ExplanationSurrogate = Union[Dict[str, fumm.Model], fumm.Model]
 ExplanationTuple = Union[Explanation,  # yapf: disable
-                         Tuple[Explanation, Dict[str, fumm.Model]]]
+                         Tuple[Explanation, ExplanationSurrogate]]
 Index = Union[int, str]
 
 
 def _input_is_valid(
         dataset: np.ndarray, predictive_model: object, as_probabilistic: bool,
-        categorical_indices: Union[None, List[Index]],
+        as_regressor: bool, categorical_indices: Union[None, List[Index]],
         class_names: Union[None, List[str]], classes_number: Union[None, int],
         feature_names: Union[None, List[str]],
         unique_predictions: Union[None, List[Union[str, int]]]) -> bool:
@@ -121,27 +122,40 @@ surrogate_explainers.SurrogateTabularExplainer` class.
 
     if not isinstance(as_probabilistic, bool):
         raise TypeError('The as_probabilistic parameter has to be a boolean.')
+    if not isinstance(as_regressor, bool):
+        raise TypeError('The as_regressor parameter has to be a boolean.')
 
-    if as_probabilistic:
-        is_functional = fumv.check_model_functionality(predictive_model, True,
-                                                       True)
-        if not is_functional:
-            raise IncompatibleModelError(
-                'With as_predictive set to True the predictive model '
-                'needs to be capable of outputting probabilities via '
-                'a *predict_proba* method, which takes exactly one '
-                'required parameter -- data to be predicted -- and '
-                'outputs a 2-dimensional array with probabilities.')
-    else:
+    if as_regressor:
         is_functional = fumv.check_model_functionality(predictive_model, False,
                                                        True)
         if not is_functional:
             raise IncompatibleModelError(
-                'With as_predictive set to False the predictive model '
-                'needs to be capable of outputting (class) predictions '
+                'With as_regressor set to True the predictive model '
+                'needs to be capable of outputting numerical predictions '
                 'via a *predict* method, which takes exactly one required '
                 'parameter -- data to be predicted -- and outputs a '
-                '1-dimensional array with (class) predictions.')
+                '1-dimensional array with numerical predictions.')
+    else:
+        if as_probabilistic:
+            is_functional = fumv.check_model_functionality(
+                predictive_model, True, True)
+            if not is_functional:
+                raise IncompatibleModelError(
+                    'With as_predictive set to True the predictive model '
+                    'needs to be capable of outputting probabilities via '
+                    'a *predict_proba* method, which takes exactly one '
+                    'required parameter -- data to be predicted -- and '
+                    'outputs a 2-dimensional array with probabilities.')
+        else:
+            is_functional = fumv.check_model_functionality(
+                predictive_model, False, True)
+            if not is_functional:
+                raise IncompatibleModelError(
+                    'With as_predictive set to False the predictive model '
+                    'needs to be capable of outputting (class) predictions '
+                    'via a *predict* method, which takes exactly one required '
+                    'parameter -- data to be predicted -- and outputs a '
+                    '1-dimensional array with (class) predictions.')
 
     if categorical_indices is not None:
         if isinstance(categorical_indices, list):
@@ -239,6 +253,10 @@ class SurrogateTabularExplainer(abc.ABC):
 
     .. versionadded:: 0.0.2
 
+    .. versionchanged:: 0.1.0
+
+       Added support for regression models.
+
     An abstract class that all surrogate explainer classes should inherit from.
     It contains an ``__init__`` method and an input validator --
     ``_explain_instance_input_is_valid`` -- for the abstract
[…]
