+++ fatf/utils/data/feature_selection/tests/test_sklearn.py
@@ -23,6 +23,7 @@ import sys
 import numpy as np
 
 import fatf
+import fatf.utils.array.validation as fuav
 import fatf.utils.data.feature_selection.sklearn as fudfs
 import fatf.utils.testing.imports as futi
 
@@ -54,6 +55,18 @@ CATEGORICAL_NP_ARRAY = np.array([
     ['a', 'f', 'c'],
     ['a', 'b', 'g']])
 # yapf: enable
+
+FEATURE_INDICES_WARNING = ('The selected number of features is larger '
+                           'than the total number of features in the '
+                           'dataset array. All of the features are being '
+                           'selected.')
+FEATURE_PERCENTAGE_LOG = (
+    'Since the number of features to be extracted was not given 24% of '
+    'features will be used. This percentage translates to 0 features, '
+    'therefore the number of features to be used is overwritten to 1. To '
+    'prevent this from happening, you should either explicitly set the '
+    'number of features via the features_number parameter or increase the '
+    'value of the features_percentage parameter.')
 
 
 def test_sklearn_import():
@@ -217,17 +230,6 @@ def test_lasso_path(caplog):
     """
     Tests :func:`fatf.utils.data.feature_choice.sklearn.lasso_path` function.
     """
-    feature_indices_warning = ('The selected number of features is larger '
-                               'than the total number of features in the '
-                               'dataset array. All of the features are being '
-                               'selected.')
-    feature_percentage_log = (
-        'Since the number of features to be extracted was not given 24% of '
-        'features will be used. This percentage translates to 0 features, '
-        'therefore the number of features to be used is overwritten to 1. To '
-        'prevent this from happening, you should either explicitly set the '
-        'number of features via the features_number parameter or increase the '
-        'value of the features_percentage parameter.')
     no_lasso_log = ('The lasso path feature selection could not pick any '
                     'feature subset. All of the features were selected.')
     less_lasso_log = ('The lasso path feature selection could not pick {} '
@@ -246,42 +248,49 @@ def test_lasso_path(caplog):
     # Classic array -- weights
     features = fudfs.lasso_path(NUMERICAL_NP_ARRAY, NUMERICAL_NP_ARRAY_TARGET,
                                 weights, 2)
+    assert fuav.is_1d_array(features)
     assert np.array_equal(features, np.array([0, 1]))
     # Structured array -- no-weights
     features = fudfs.lasso_path(
         NUMERICAL_STRUCT_ARRAY, NUMERICAL_NP_ARRAY_TARGET, features_number=2)
+    assert fuav.is_1d_array(features)
     assert np.array_equal(features, np.array(['a', 'b']))
     #
     # Selecting exactly 4 features -- no need for Lasso
     features = fudfs.lasso_path(NUMERICAL_NP_ARRAY, NUMERICAL_NP_ARRAY_TARGET,
                                 weights, 4)
+    assert fuav.is_1d_array(features)
     assert np.array_equal(features, np.array([0, 1, 2, 3]))
     # Selecting more than 4 features
     with pytest.warns(UserWarning) as warning:
         features = fudfs.lasso_path(NUMERICAL_STRUCT_ARRAY,
[…]
