+++ fatf/utils/data/feature_selection/sklearn.py
@@ -8,7 +8,7 @@ scikit-learn-based feature selection approaches.
 #         Kacper Sokol <k.sokol@bristol.ac.uk>
 # License: new BSD
 
-from typing import List, Union, Optional
+from typing import List, Optional, Union
 
 import logging
 import warnings
@@ -31,9 +31,7 @@ import fatf.utils.array.validation as fuav
 
 from fatf.exceptions import IncorrectShapeError
 
-__all__ = ['lasso_path']
-
-Index = Union[int, str]
+__all__ = ['lasso_path', 'forward_selection', 'highest_weights']
 
 logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
 
@@ -105,11 +103,35 @@ def _validate_input_lasso_path(dataset: np.ndarray, target: np.ndarray,
     return input_is_valid
 
 
+def _get_feature_proportion(features_percentage: int,
+                            indices_number: int) -> int:
+    """
+    Computes a number of features based on the given percentage.
+    """
+    assert (isinstance(features_percentage, int)
+            and 0 <= features_percentage <= 100
+            and isinstance(indices_number, int))
+    feature_proportion = int((features_percentage / 100) * indices_number)
+    if feature_proportion:
+        features_number = feature_proportion
+    else:
+        logger.warning(
+            'Since the number of features to be extracted was not given '
+            '%d%% of features will be used. This percentage translates to '
+            '0 features, therefore the number of features to be used is '
+            'overwritten to 1. To prevent this from happening, you should '
+            'either explicitly set the number of features via the '
+            'features_number parameter or increase the value of the '
+            'features_percentage parameter.', features_percentage)
+        features_number = feature_proportion + 1
+    return features_number
+
+
 def lasso_path(dataset: np.ndarray,
                target: np.ndarray,
                weights: Optional[np.ndarray] = None,
                features_number: Optional[int] = None,
-               features_percentage: int = 100) -> List[Index]:
+               features_percentage: int = 100) -> np.ndarray:
     """
     Selects the specified number of features based on Lasso path coefficients.
 
@@ -131,34 +153,32 @@ def lasso_path(dataset: np.ndarray,
     weighted) and scaling by the square root of the ``weights``.
     The ``target`` array is treated in the same way.
 
-    This feature selection method is based on the default feature selection
-    mechanism implemented by LIME_ (Local Interpretable Model-agnostic
-    Explanations. The original implementation can be found in the
-    ``lime.lime_base.LimeBase.feature_selection`` method in the
+    This feature selection method is based on LIME_ (Local Interpretable
+    Model-agnostic Explanations). The original implementation can be found in
+    the ``lime.lime_base.LimeBase.feature_selection`` method in the
     `official LIME package`_.
 
     .. _LIME: https://github.com/marcotcr/lime
-    .. _`official LIME package`: https://github.com/marcotcr/lime/blob/master/
-       lime/lime_base.py#L116
+    .. _`official LIME package`: https://github.com/marcotcr/lime/blob/0.2.0.0/
+       lime/lime_base.py#L115
 
     Parameters
     ----------
     dataset : numpy.ndarray
-        A 2-dimensional numpy array with holding a data set.
+        A 2-dimensional numpy array holding a data set.
     target : numpy.ndarray
-        The class/probabilities/regression values of each row in the input data
+        The class/probability/regression values of each row in the input data
         set.
     weights : numpy.ndarray, optional (default=None)
         An array of (importance) weights for each data point in the input data
-        set. If ``None``, all of the data points are the same important when
+        set. If ``None``, all of the data points are equally important when
         computing the Lasso path.
     features_number : integer, optional (default=None)
         The number of (top) features to be selected. If ``None``, the top x% of
[…]
