+++ fatf/transparency/predictions/tests/test_surrogate_explainers.py
@@ -860,6 +860,12 @@ TabularBlimeyLime` class.
             numerical_np_array_classifier,
             as_regressor=True)
 
+        wide_data = np.concatenate(2 * [futt.NUMERICAL_NP_ARRAY], axis=1)
+        numerical_np_array_classifier_wide = fum.KNN(k=3)
+        numerical_np_array_classifier_wide.fit(wide_data, futt.LABELS)
+        numerical_np_tabular_lime_wide_reg = ftps.TabularBlimeyLime(
+            wide_data, numerical_np_array_classifier_wide, as_regressor=True)
+
         numerical_struct_array_classifier = fum.KNN(k=3)
         numerical_struct_array_classifier.fit(futt.NUMERICAL_STRUCT_ARRAY,
                                               futt.LABELS)
@@ -1131,14 +1137,26 @@ predictions.surrogate_explainers.TabularBlimeyLime.explain_instance` method.
                 futt.NUMERICAL_NP_ARRAY[0], explained_class=3)
         assert str(exin.value) == explain_class_value_error2.format(2, 3)
 
-    def test_explain_instance(self):
+    def test_explain_instance(self, caplog):
         """
         Tests the ``explain_instance`` method.
 
         Tests :func:`fatf.transparency.predictions.surrogate_explainers.\
 TabularBlimeyLime.explain_instance` method.
         """
+        assert len(caplog.records) == 0
         fatf.setup_random_seed()
+        assert len(caplog.records) == 2
+        assert caplog.records[0].levelname == 'INFO'
+        assert caplog.records[0].getMessage() == ('Seeding RNGs using the '
+                                                  'system variable.')
+        assert caplog.records[1].levelname == 'INFO'
+        assert caplog.records[1].getMessage() == 'Seeding RNGs with 42.'
+
+        log_info_forward_selection = ('Selecting {} features with forward '
+                                      'selection.')
+        log_info_highest_weights = ('Selecting {} features with highest '
+                                    'weights.')
 
         numerical_np_explanation = {
             'class 0': {
@@ -1165,6 +1183,25 @@ TabularBlimeyLime.explain_instance` method.
             '*feature 1* <= 0.00': -0.659,
             '0.07 < *feature 2* <= 0.22': -0.065,
             '0.58 < *feature 3* <= 0.79': -0.015
+        }
+        numerical_np_explanation_wide_reg = {
+            '*feature 0* <= 0.00': 0.571,
+            '*feature 1* <= 0.00': -0.626,
+            '0.07 < *feature 2* <= 0.22': -0.328,
+            '*feature 4* <= 0.00': 0.442,
+            '*feature 5* <= 0.00': -0.265,
+            '0.07 < *feature 6* <= 0.22': 0.371,
+            '0.58 < *feature 7* <= 0.79': 0.230
+        }
+        numerical_np_explanation_wide_reg_all = {
+            '*feature 0* <= 0.00': 0.752,
+            '*feature 1* <= 0.00': -0.260,
+            '0.07 < *feature 2* <= 0.22': 0.246,
+            '0.58 < *feature 3* <= 0.79': -0.464,
+            '*feature 4* <= 0.00': 0.811,
+            '*feature 5* <= 0.00': -0.305,
+            '0.07 < *feature 6* <= 0.22': -0.036,
+            '0.58 < *feature 7* <= 0.79': 0.096
         }
         numerical_struct_explanation = {
             'class 0': {
@@ -1209,6 +1246,7 @@ TabularBlimeyLime.explain_instance` method.
             }
         }
 
+        assert len(caplog.records) == 2
         # Probabilistic
[…]
