### 2020-05-18 04dd83e
Documentation update and fatf version bump to 0.1.0
