+++ doc/news.rst
@@ -5,13 +5,18 @@
 News
 ++++
 
-[01/08/2019] Version 0.0.1 released!
+[18/05/2020] Version 0.1.0 released!
 ====================================
 
-Today we release our first public version of the FAT Forensics package. To
-familiarise yourself with the FAT Forensics Python package make sure to check
-out our :ref:`getting_started` page. The changelog summarises functionality
-of this release: :ref:`changelog_0_0_1`.
+Today we release an incremental update focused on surrogate explainers.
+This release coincides with publication of a `paper <JOSS_paper_>`_
+describing FAT Forensic in The Journal of Open Source Software (JOSS).
+The changelog summarises the functionality added with this release:
+:ref:`changelog_0_1_0`.
+
+Next up, surrogate explainability for image and text data.
+
+.. _JOSS_paper: https://joss.theoj.org/papers/070c8b6b705bb47d1432673a1eb03f0c
 
 [04/11/2019] Version 0.0.2 released!
 ====================================
@@ -32,3 +37,11 @@ instead. The ":ref:`tutorials_prediction_explainability`" tutorial has also
 been updated to use bLIMEy instead of LIME.
 
 Surrogate explainability for image and text data is coming soon so stay tuned.
+
+[01/08/2019] Version 0.0.1 released!
+====================================
+
+Today we release our first public version of the FAT Forensics package. To
+familiarise yourself with the FAT Forensics Python package make sure to check
+out our :ref:`getting_started` page. The changelog summarises functionality
+of this release: :ref:`changelog_0_0_1`.
