+++ examples/fairness/README.txt
@@ -3,6 +3,6 @@
 Fairness Examples
 =================
 
-Here you can find various examples of how to use FAT-Forensics to evaluate and
+Here you can find various examples of how to use FAT Forensics to evaluate and
 mitigate a range of fairness aspects of data sets, machine learning models and
 their predictions.

[…]
