### 2020-05-18 8cf1d57
0.1.0 release and JOSS publication (#32)
* scikit-learn>=0.22 compatibility and general improvements
* scikit-learn>=0.23 and numpy>=1.18 compatibility and testing
* ImportWarning -> ImportError
* LIME wrapper deprecation
* LIME visualisation cleanup
* scikit-learn .postX version fix
* scikit-learn -- numpy and scipy compatibility fix
* scikit-learn>=0.22 pylint fix when an older version is installed
* Black-box regression support for surrogates
* Implementation of LIME's forward selection and highest weight feature selection
* Correct feature selection algorithms in LIME
* Documentation updat
[…]
