+++ fatf/transparency/predictions/surrogate_explainers.py
@@ -45,6 +45,7 @@ import fatf.utils.distances as fud
 import fatf.utils.kernels as fatf_kernels
 import fatf.utils.models.validation as fumv
 import fatf.utils.models.models as fumm
+import fatf.utils.tools as fut
 
 __all__ = ['SurrogateTabularExplainer',
            'TabularBlimeyLime',
@@ -54,11 +55,15 @@ logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
 
 try:
     # pylint: disable=ungrouped-imports
+    import sklearn
     import sklearn.linear_model
     import sklearn.tree
 
     import fatf.transparency.sklearn.linear_model as ftslm
     import fatf.utils.data.feature_selection.sklearn as fudfs
+
+    _SKLEARN_VERSION = [int(i) for i in sklearn.__version__.split('.')[:2]]
+    _SKLEARN_0_22 = fut.at_least_verion([0, 22], _SKLEARN_VERSION)
 except ImportError as exin:
     warnings.warn(
         'The TabularBlimeyLime and TabularBlimeyTree surrogate explainers '
@@ -69,20 +74,26 @@ except ImportError as exin:
     ReturnTree = None  # pylint: disable=invalid-name
     SKLEARN_MISSING = True
 else:
-    ReturnTree = sklearn.tree.tree.BaseDecisionTree
+    if _SKLEARN_0_22:  # pragma: nocover
+        # pylint: disable=invalid-name,protected-access,no-member
+        ReturnTree = sklearn.tree._classes.BaseDecisionTree
+    else:  # pragma: nocover
+        ReturnTree = (  # pylint: disable=invalid-name
+            sklearn.tree.tree.BaseDecisionTree)
     SKLEARN_MISSING = False
 
 BinSamplingValues = Dict[Union[str, int],
                          Dict[int, Tuple[float, float, float, float]]]
 Explanation = Dict[str, Dict[str, float]]
+ExplanationSurrogate = Union[Dict[str, fumm.Model], fumm.Model]
 ExplanationTuple = Union[Explanation,  # yapf: disable
-                         Tuple[Explanation, Dict[str, fumm.Model]]]
+                         Tuple[Explanation, ExplanationSurrogate]]
 Index = Union[int, str]
 
 
 def _input_is_valid(
         dataset: np.ndarray, predictive_model: object, as_probabilistic: bool,
-        categorical_indices: Union[None, List[Index]],
+        as_regressor: bool, categorical_indices: Union[None, List[Index]],
         class_names: Union[None, List[str]], classes_number: Union[None, int],
         feature_names: Union[None, List[str]],
         unique_predictions: Union[None, List[Union[str, int]]]) -> bool:
@@ -111,27 +122,40 @@ surrogate_explainers.SurrogateTabularExplainer` class.
 
     if not isinstance(as_probabilistic, bool):
         raise TypeError('The as_probabilistic parameter has to be a boolean.')
+    if not isinstance(as_regressor, bool):
+        raise TypeError('The as_regressor parameter has to be a boolean.')
 
-    if as_probabilistic:
-        is_functional = fumv.check_model_functionality(predictive_model, True,
-                                                       True)
-        if not is_functional:
-            raise IncompatibleModelError(
-                'With as_predictive set to True the predictive model '
-                'needs to be capable of outputting probabilities via '
-                'a *predict_proba* method, which takes exactly one '
-                'required parameter -- data to be predicted -- and '
-                'outputs a 2-dimensional array with probabilities.')
-    else:
+    if as_regressor:
         is_functional = fumv.check_model_functionality(predictive_model, False,
                                                        True)
         if not is_functional:
             raise IncompatibleModelError(
-                'With as_predictive set to False the predictive model '
-                'needs to be capable of outputting (class) predictions '
[…]
