### 2020-05-18 c723a6f
Travis CI config update and fix (matrix and jobs conflict)
