+++ .travis.yml
@@ -1,5 +1,6 @@
 # Use container on EC2, other options available
 # (venvs)[https://docs.travis-ci.com/user/reference/overview/]
+os: linux
 dist: xenial
 
 notifications:
@@ -29,7 +30,7 @@ python:
 env:
   global:
     - MAKEFLAGS='-j 2'  # This allows to parallelise make, e.g. make t1 t2
-  matrix:
+  jobs:
     - FATF_TEST_SCIPY=1.4.1  FATF_TEST_NUMPY=1.18.4
     - FATF_TEST_SCIPY=1.3.3  FATF_TEST_NUMPY=1.17.5
     - FATF_TEST_SCIPY=1.3.0  FATF_TEST_NUMPY=1.16.3
@@ -43,7 +44,20 @@ env:
     - FATF_TEST_SCIPY=0.18.1 FATF_TEST_NUMPY=1.10.1 FATF_TEST_SKLEARN=0.20.4 FATF_TEST_MATPLOTLIB=3.0.0
     - FATF_TEST_SCIPY=latest FATF_TEST_NUMPY=latest
 
-matrix:
+install:
+  - make install-matplotlib
+  - make install-scikit-learn
+  - make dependencies-dev  # Install python development dependencies
+
+script:
+  - make linting-pylint
+  - make linting-flake8
+  - make linting-yapf
+  - make check-types
+  - make test-with-code-coverage
+#  - make test-notebooks
+
+jobs:
   fast_finish: true
   exclude:
     - python: '3.7'
@@ -79,21 +93,6 @@ matrix:
   allow_failures:
     - python: '3.9-dev'
     - env: FATF_TEST_SCIPY=latest FATF_TEST_NUMPY=latest
-
-install:
-  - make install-matplotlib
-  - make install-scikit-learn
-  - make dependencies-dev  # Install python development dependencies
-
-script:
-  - make linting-pylint
-  - make linting-flake8
-  - make linting-yapf
-  - make check-types
-  - make test-with-code-coverage
-#  - make test-notebooks
-
-jobs:
   include:
     - stage: Documentation Testing and Deployment
       python: 3.6
@@ -109,7 +108,7 @@ jobs:
       deploy:
         provider: script
         script: bash build_tools/doc/deploy-doc.sh ${TRAVIS_BRANCH}
-        skip_cleanup: true
+        cleanup: false
         on: &deploy-conditions
           repo: fat-forensics/fat-forensics
           branch: master
@@ -122,7 +121,7 @@ jobs:
       deploy:
         provider: script
         script: make deploy-code-coverage
-        skip_cleanup: true
+        cleanup: false
         on:
           <<: *deploy-conditions
           all_branches: true
@@ -141,7 +140,7 @@ jobs:
[…]
