### 2020-05-18 1aa74c0
Travis CI jobs env fix
