+++ .travis.yml
@@ -96,9 +96,7 @@ jobs:
   include:
     - stage: Documentation Testing and Deployment
       python: 3.6
-      if: |
-        env(FATF_TEST_SCIPY) = 1.3.0 AND \
-        env(FATF_TEST_NUMPY) = 1.16.3
+      env: FATF_TEST_SCIPY=1.3.0  FATF_TEST_NUMPY=1.16.3
       script:
         - make test-doc
         - make doc-linkcheck
@@ -114,9 +112,7 @@ jobs:
           branch: master
     - stage: Deploy Code Coverage
       python: 3.6
-      if: |
-        env(FATF_TEST_SCIPY) = 1.3.0 AND \
-        env(FATF_TEST_NUMPY) = 1.16.3
+      env: FATF_TEST_SCIPY=1.3.0  FATF_TEST_NUMPY=1.16.3
       script: make test-with-code-coverage
       deploy:
         provider: script
@@ -127,15 +123,13 @@ jobs:
           all_branches: true
     - stage: Release on PyPi
       python: 3.6
-      if: |
-        env(FATF_TEST_SCIPY) = 1.3.0 AND \
-        env(FATF_TEST_NUMPY) = 1.16.3
+      env: FATF_TEST_SCIPY=1.3.0  FATF_TEST_NUMPY=1.16.3
       install: make install-sans-dep
       script: skip
       before_deploy: bash build_tools/pypi/version-validation.sh
       deploy:
         provider: pypi
-        user: "__token__"
+        username: "__token__"
         password:
           secure: "OC0Z4KEyI27x9p2PeohulMh9SWWBGwX2EgK5+rIkjM8Crr/IVZC2UmJnwi4CC2xQHYbB5qRLEQo4lzHXPKiNgEoHubUZo1ZKnlqJI+feyDSvlPRoBrm5aTdWj7sI30kTXYbAYFN1roE+PA1ZVhJ9206stq+BF5MvGoVloy+DuGP3xEMI3iB/AQd4Fu5OKrM0z5mTJMSd2m+mo47gQ+NoyGXNWJmAkMOKA5p9zqpcS9SgSuBUgIY9vAg+huC4Kn56cJRKe8aZlpTyLEYnggvY63MAlaNF2lY006/clwRUYgNh56fejecJFMdcdCqwPTSJiPfJDQsnwqqQ3QW8NlPFne2Ct7nKFWrpmC9NkbkLRSapcADmeTBpucXZt2F78wS1yagCLcCVYRRhs0bWswmbNVSERuaNo/nZ3kiOrSwFV3/foHqBwVYUKgmyCxPObf+iYWslPLpam/RuHdeoWaOhvKdl/xgvJw/qUCJoTBC2W/Z+OVjNPYmk/PtzcQ8CZ3jRfYg0pyr9zFSjMfGCVRpVU7or4muOC9MtVH8eeGsZyVj3CgWc85/JF/G6MHF0OuJDbBsWHsSekLKSjCJy3aCpJSyuOw+VHyUDyTslCLAc5Lh5PQkVK23lYjrDWVjSPD3Y6pFxQLq+SI3hLfvMoZvtfx9QHFPq++Gh9FQMixFtYX0="
         distributions: "sdist bdist_wheel"
@@ -147,9 +141,7 @@ jobs:
           condition: "$TRAVIS_TAG =~ ^[0-9]+\\.[0-9]+\\.[0-9]+(-rev[0-9]+)?$"
     - stage: Release on GitHub
       python: 3.6
-      if: |
-        env(FATF_TEST_SCIPY) = 1.3.0 AND \
[…]
