### 2020-05-19 #1904 @So-Cool
Hi @arfon and @arokem,
apologies for the delay. The release tag is [`0.1.0`](https://github.com/fat-forensics/fat-forensics/releases/tag/0.1.0). The DOI of the release is [10.5281/zenodo.3833199](https://doi.org/10.5281/zenodo.3833199).
